@@ -18,7 +18,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 2.2.3
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
+  sdk: ">=3.8.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -61,11 +61,11 @@ dependencies:
     path: ./external_path
   filesize: ^2.0.1
   fluttericon: ^2.0.0
-  fluttertoast: ^8.2.4
+  fluttertoast: ^9.0.0
   flutter_cache_manager: ^3.3.1
   flutter_displaymode: ^0.6.0
   #flutter_file_dialog: ^3.0.1
-  flutter_inappwebview: ^6.0.0
+  flutter_inappwebview: ^6.1.5
   flutter_local_notifications: ^17.0.0
   flutter_material_color_picker: ^1.2.0
   flutter_screenutil: ^5.9.0
@@ -73,8 +73,8 @@ dependencies:
   google_fonts: ^6.1.0
   html: ^0.15.3
   http: ^1.2.1
-  i18n_extension: ^12.0.1
-  intl: ^0.19.0
+  i18n_extension: ^15.0.5
+  intl: ^0.20.2
   json_annotation: ^4.9.0
   just_audio: ^0.9.39
   #just_audio:
@@ -115,21 +115,16 @@ dependencies:
   rxdart: ^0.28.0
   scrobblenaut:
     path: ./scrobblenaut
-  share_plus: ^9.0.0
+  share_plus: ^12.0.0
   spotify: ^0.13.1
-  sqflite: ^2.3.0
+  sqflite: ^2.4.2
   url_launcher: ^6.2.2
   #version: ^3.0.2 Added contents in source directly for compatibility
   visibility_detector: ^0.4.0+2
   wakelock_plus: ^1.2.4
 
-dependency_overrides:
-  # TODO: Recheck once flutter_inappwebview version >6.0.0 is released
-  flutter_inappwebview_android:
-    git:
-      url: https://github.com/holzgeist/flutter_inappwebview
-      path: flutter_inappwebview_android
-      ref: d89b1d32638b49dfc58c4b7c84153be0c269d057
+dependency_overrides: 
+  sqflite_android: 2.4.1 
 
 dev_dependencies:
   flutter_test:
