@@ -24,8 +24,7 @@ DownloadManager downloadManager = DownloadManager();
 class DownloadManager {
   //Platform channels
   static const MethodChannel platform = MethodChannel('s.s.saturn/native');
-  static const EventChannel eventChannel =
-      EventChannel('s.s.saturn/downloads');
+  static const EventChannel eventChannel = EventChannel('s.s.saturn/downloads');
 
   bool running = false;
   int queueSize = 0;
@@ -171,11 +170,21 @@ class DownloadManager {
       builder: (context) {
         return AlertDialog(
           title: Text('Storage Permission Required'.i18n),
-          content: Text('Storage permission is required to download content.\nPlease open system settings and grant storage permission to Saturn.'.i18n),
+          content: Text(
+              'Storage permission is required to download content.\nPlease open system settings and grant storage permission to Saturn.'
+                  .i18n),
           actions: <Widget>[
             TextButton(
               style: ButtonStyle(
-                overlayColor: WidgetStateProperty.resolveWith<Color?>((Set<WidgetState> states) {if (states.contains(WidgetState.pressed)) {return Theme.of(context).primaryColor.withOpacity(0.3);}return null;}),
+                overlayColor: WidgetStateProperty.resolveWith<Color?>(
+                    (Set<WidgetState> states) {
+                  if (states.contains(WidgetState.pressed)) {
+                    return Theme.of(context)
+                        .primaryColor
+                        .withValues(alpha: 0.3);
+                  }
+                  return null;
+                }),
               ),
               child: Text('Cancel'.i18n),
               onPressed: () {
@@ -185,7 +194,15 @@ class DownloadManager {
             ),
             TextButton(
               style: ButtonStyle(
-                overlayColor: WidgetStateProperty.resolveWith<Color?>((Set<WidgetState> states) {if (states.contains(WidgetState.pressed)) {return Theme.of(context).primaryColor.withOpacity(0.3);}return null;}),
+                overlayColor: WidgetStateProperty.resolveWith<Color?>(
+                    (Set<WidgetState> states) {
+                  if (states.contains(WidgetState.pressed)) {
+                    return Theme.of(context)
+                        .primaryColor
+                        .withValues(alpha: 0.3);
+                  }
+                  return null;
+                }),
               ),
               child: Text('Open system settings'.i18n),
               onPressed: () {
@@ -215,7 +232,15 @@ class DownloadManager {
           actions: <Widget>[
             TextButton(
               style: ButtonStyle(
-                overlayColor: WidgetStateProperty.resolveWith<Color?>((Set<WidgetState> states) {if (states.contains(WidgetState.pressed)) {return Theme.of(context).primaryColor.withOpacity(0.3);}return null;}),
+                overlayColor: WidgetStateProperty.resolveWith<Color?>(
+                    (Set<WidgetState> states) {
+                  if (states.contains(WidgetState.pressed)) {
+                    return Theme.of(context)
+                        .primaryColor
+                        .withValues(alpha: 0.3);
+                  }
+                  return null;
+                }),
               ),
               child: Text('Cancel'.i18n),
               onPressed: () {
@@ -225,7 +250,15 @@ class DownloadManager {
             ),
             TextButton(
               style: ButtonStyle(
-                overlayColor: WidgetStateProperty.resolveWith<Color?>((Set<WidgetState> states) {if (states.contains(WidgetState.pressed)) {return Theme.of(context).primaryColor.withOpacity(0.3);}return null;}),
+                overlayColor: WidgetStateProperty.resolveWith<Color?>(
+                    (Set<WidgetState> states) {
+                  if (states.contains(WidgetState.pressed)) {
+                    return Theme.of(context)
+                        .primaryColor
+                        .withValues(alpha: 0.3);
+                  }
+                  return null;
+                }),
               ),
               child: Text('Continue'.i18n),
               onPressed: () {
