@@ -1,3 +1,5 @@
+// ignore_for_file: use_build_context_synchronously
+
 import 'dart:async';
 
 import 'package:flutter/material.dart';
@@ -19,15 +21,12 @@ import '../ui/cached_image.dart';
 import '../ui/details_screens.dart';
 import '../ui/error.dart';
 
-import '../api/clubs.dart';
-ClubRoom clubroom = ClubRoom();
-
 class MenuSheet {
   Function navigateCallback;
 
   // Use no-op callback if not provided
   MenuSheet({Function? navigateCallback})
-      : navigateCallback = navigateCallback ?? (() {});
+    : navigateCallback = navigateCallback ?? (() {});
 
   //===================
   // DEFAULT
@@ -35,21 +34,20 @@ class MenuSheet {
 
   void show(BuildContext context, List<Widget> options) {
     showModalBottomSheet(
-        isScrollControlled: true,
-        context: context,
-        builder: (BuildContext context) {
-          return ConstrainedBox(
-            constraints: BoxConstraints(
-              maxHeight:
-                  (MediaQuery.of(context).orientation == Orientation.landscape)
-                      ? 220
-                      : 350,
-            ),
-            child: SingleChildScrollView(
-              child: Column(children: options),
-            ),
-          );
-        });
+      isScrollControlled: true,
+      context: context,
+      builder: (BuildContext context) {
+        return ConstrainedBox(
+          constraints: BoxConstraints(
+            maxHeight:
+                (MediaQuery.of(context).orientation == Orientation.landscape)
+                ? 220
+                : 350,
+          ),
+          child: SingleChildScrollView(child: Column(children: options)),
+        );
+      },
+    );
   }
 
   //===================
@@ -58,96 +56,86 @@ class MenuSheet {
 
   void showWithTrack(BuildContext context, Track track, List<Widget> options) {
     showModalBottomSheet(
-        isScrollControlled: true,
-        context: context,
-        builder: (BuildContext context) {
-          return Column(
-            mainAxisSize: MainAxisSize.min,
-            children: <Widget>[
-              Container(
-                height: 16.0,
-              ),
-              Row(
-                mainAxisSize: MainAxisSize.max,
-                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                children: <Widget>[
-                  Semantics(
-                    label: 'Album art'.i18n,
-                    image: true,
-                    child: CachedImage(
-                      url: track.albumArt?.full ?? '',
-                      height: 128,
-                      width: 128,
-                    ),
+      isScrollControlled: true,
+      context: context,
+      builder: (BuildContext context) {
+        return Column(
+          mainAxisSize: MainAxisSize.min,
+          children: <Widget>[
+            Container(height: 16.0),
+            Row(
+              mainAxisSize: MainAxisSize.max,
+              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+              children: <Widget>[
+                Semantics(
+                  label: 'Album art'.i18n,
+                  image: true,
+                  child: CachedImage(
+                    url: track.albumArt?.full ?? '',
+                    height: 128,
+                    width: 128,
                   ),
-                  SizedBox(
-                    width: 240.0,
-                    child: Column(
-                      mainAxisSize: MainAxisSize.min,
-                      children: <Widget>[
-                        Text(
-                          track.title ?? '',
-                          maxLines: 1,
-                          textAlign: TextAlign.center,
-                          overflow: TextOverflow.ellipsis,
-                          style: const TextStyle(
-                              fontSize: 22.0, fontWeight: FontWeight.bold),
+                ),
+                SizedBox(
+                  width: 240.0,
+                  child: Column(
+                    mainAxisSize: MainAxisSize.min,
+                    children: <Widget>[
+                      Text(
+                        track.title ?? '',
+                        maxLines: 1,
+                        textAlign: TextAlign.center,
+                        overflow: TextOverflow.ellipsis,
+                        style: const TextStyle(
+                          fontSize: 22.0,
+                          fontWeight: FontWeight.bold,
                         ),
-                        Text(
-                          track.artistString ?? '',
-                          textAlign: TextAlign.center,
-                          overflow: TextOverflow.ellipsis,
-                          maxLines: 1,
-                          style: const TextStyle(fontSize: 20.0),
-                        ),
-                        Container(
-                          height: 8.0,
-                        ),
-                        Text(
-                          track.album?.title ?? '',
-                          textAlign: TextAlign.center,
-                          overflow: TextOverflow.ellipsis,
-                          maxLines: 1,
-                        ),
-                        Text(track.durationString ?? '')
-                      ],
-                    ),
+                      ),
+                      Text(
+                        track.artistString ?? '',
+                        textAlign: TextAlign.center,
+                        overflow: TextOverflow.ellipsis,
+                        maxLines: 1,
+                        style: const TextStyle(fontSize: 20.0),
+                      ),
+                      Container(height: 8.0),
+                      Text(
+                        track.album?.title ?? '',
+                        textAlign: TextAlign.center,
+                        overflow: TextOverflow.ellipsis,
+                        maxLines: 1,
+                      ),
+                      Text(track.durationString ?? ''),
+                    ],
                   ),
-                ],
-              ),
-              Container(
-                height: 16.0,
-              ),
-              ConstrainedBox(
-                constraints: BoxConstraints(
-                  maxHeight: (MediaQuery.of(context).orientation ==
-                          Orientation.landscape)
-                      ? 200
-                      : 350,
                 ),
-                child: SingleChildScrollView(
-                  child: Column(children: options),
-                ),
-              )
-            ],
-          );
-        });
+              ],
+            ),
+            Container(height: 16.0),
+            ConstrainedBox(
+              constraints: BoxConstraints(
+                maxHeight:
+                    (MediaQuery.of(context).orientation ==
+                        Orientation.landscape)
+                    ? 200
+                    : 350,
+              ),
+              child: SingleChildScrollView(child: Column(children: options)),
+            ),
+          ],
+        );
+      },
+    );
   }
 
   //Default track options
-  void defaultTrackMenu(Track track,
-      {required BuildContext context,
-      List<Widget> options = const [],
-      Function? onRemove}) {
-
-      List<Widget> queueOptions = [];
-
-      if (clubroom.ifhost()) {
-        queueOptions.add(addToQueueNext(track, context));
-        queueOptions.add(addToQueue(track, context));
-      } else {
-        queueOptions.add(requestSong(track, context));
-      }
+  void defaultTrackMenu(
+    Track track, {
+    required BuildContext context,
+    List<Widget> options = const [],
+    Function? onRemove,
+  }) {
+    List<Widget> queueOptions = [];
 
     showWithTrack(context, track, [
       ...queueOptions,
@@ -160,113 +148,104 @@ class MenuSheet {
       shareTile('track', track.id ?? ''),
       playMix(track, context),
       showAlbum(track.album!, context),
-      ...List.generate(track.artists?.length ?? 0,
-          (i) => showArtist(track.artists![i], context)),
-      ...options
+      ...List.generate(
+        track.artists?.length ?? 0,
+        (i) => showArtist(track.artists![i], context),
+      ),
+      ...options,
     ]);
   }
 
   //===================
   // TRACK OPTIONS
   //===================
-    
-    dead() async { 
-    if (clubroom.ifhost()) {
-      return();
-    } else {
-      return();
-    }}
-
-  Widget requestSong(Track t, BuildContext context) => ListTile(
-      title: Text('Request Song'.i18n),
-      leading: const Icon(Icons.music_note),
-      onTap: () async {
-        ClubRoom clubRoom = ClubRoom();
-        SocketManagement socketManagement = SocketManagement(address: 'https://clubs.saturn.kim:443', clubRoom: clubRoom);
-        print('precall');
-        socketManagement.songRequest(t.id.toString());
-        print('aftercall');
-        if (context.mounted) _close(context);
-      });
 
   Widget addToQueueNext(Track t, BuildContext context) => ListTile(
-      title: Text('Play next'.i18n),
-      leading: const Icon(Icons.playlist_play),
-      onTap: () async {
-        //-1 = next
-        await GetIt.I<AudioPlayerHandler>()
-            .insertQueueItem(-1, t.toMediaItem());
-        if (context.mounted) _close(context);
-      });
+    title: Text('Play next'.i18n),
+    leading: const Icon(Icons.playlist_play),
+    onTap: () async {
+      //-1 = next
+      await GetIt.I<AudioPlayerHandler>().insertQueueItem(-1, t.toMediaItem());
+      if (context.mounted) _close(context);
+    },
+  );
 
   Widget addToQueue(Track t, BuildContext context) => ListTile(
-      title: Text('Add to queue'.i18n),
-      leading: const Icon(Icons.playlist_add),
-      onTap: () async {
-        await GetIt.I<AudioPlayerHandler>().addQueueItem(t.toMediaItem());
-        if (context.mounted) _close(context);
-      });
+    title: Text('Add to queue'.i18n),
+    leading: const Icon(Icons.playlist_add),
+    onTap: () async {
+      await GetIt.I<AudioPlayerHandler>().addQueueItem(t.toMediaItem());
+      if (context.mounted) _close(context);
+    },
+  );
 
   Widget addTrackFavorite(Track t, BuildContext context) => ListTile(
-      title: Text('Add track to favorites'.i18n),
-      leading: const Icon(Icons.favorite),
-      onTap: () async {
-        await deezerAPI.addFavoriteTrack(t.id!);
-        //Make track offline, if favorites are offline
-        Playlist p = Playlist(id: deezerAPI.favoritesPlaylistId);
-        if (await downloadManager.checkOffline(playlist: p)) {
-          downloadManager.addOfflinePlaylist(p);
-        }
-        Fluttertoast.showToast(
-            msg: 'Added to library'.i18n,
-            gravity: ToastGravity.BOTTOM,
-            toastLength: Toast.LENGTH_SHORT);
-        //Add to cache
-        cache.libraryTracks ??= [];
-        cache.libraryTracks?.add(t.id!);
+    title: Text('Add track to favorites'.i18n),
+    leading: const Icon(Icons.favorite),
+    onTap: () async {
+      await deezerAPI.addFavoriteTrack(t.id!);
+      //Make track offline, if favorites are offline
+      Playlist p = Playlist(id: deezerAPI.favoritesPlaylistId);
+      if (await downloadManager.checkOffline(playlist: p)) {
+        downloadManager.addOfflinePlaylist(p);
+      }
+      Fluttertoast.showToast(
+        msg: 'Added to library'.i18n,
+        gravity: ToastGravity.BOTTOM,
+        toastLength: Toast.LENGTH_SHORT,
+      );
+      //Add to cache
+      cache.libraryTracks ??= [];
+      cache.libraryTracks?.add(t.id!);
 
-        if (context.mounted) _close(context);
-      });
+      if (context.mounted) _close(context);
+    },
+  );
 
   Widget downloadTrack(Track t, BuildContext context) => ListTile(
-        title: Text('Download'.i18n),
-        leading: const Icon(Icons.file_download),
-        onTap: () async {
-          if (await downloadManager.addOfflineTrack(t,
-                  private: false, isSingleton: true) !=
-              false) {
-            showDownloadStartedToast();
-          }
-          if (context.mounted) _close(context);
-        },
-      );
+    title: Text('Download'.i18n),
+    leading: const Icon(Icons.file_download),
+    onTap: () async {
+      if (await downloadManager.addOfflineTrack(
+            t,
+            private: false,
+            isSingleton: true,
+          ) !=
+          false) {
+        showDownloadStartedToast();
+      }
+      if (context.mounted) _close(context);
+    },
+  );
 
   Widget addToPlaylist(Track t, BuildContext context) => ListTile(
-        title: Text('Add to playlist'.i18n),
-        leading: const Icon(Icons.playlist_add),
-        onTap: () async {
-          //Show dialog to pick playlist
-          await showDialog(
-              context: context,
-              builder: (context) {
-                return SelectPlaylistDialog(
-                    track: t,
-                    callback: (Playlist p) async {
-                      await deezerAPI.addToPlaylist(t.id!, p.id!);
-                      //Update the playlist if offline
-                      if (await downloadManager.checkOffline(playlist: p)) {
-                        downloadManager.addOfflinePlaylist(p);
-                      }
-                      Fluttertoast.showToast(
-                        msg: 'Track added to'.i18n + ' ${p.title}',
-                        toastLength: Toast.LENGTH_SHORT,
-                        gravity: ToastGravity.BOTTOM,
-                      );
-                    });
-              });
-          if (context.mounted) _close(context);
+    title: Text('Add to playlist'.i18n),
+    leading: const Icon(Icons.playlist_add),
+    onTap: () async {
+      //Show dialog to pick playlist
+      await showDialog(
+        context: context,
+        builder: (context) {
+          return SelectPlaylistDialog(
+            track: t,
+            callback: (Playlist p) async {
+              await deezerAPI.addToPlaylist(t.id!, p.id!);
+              //Update the playlist if offline
+              if (await downloadManager.checkOffline(playlist: p)) {
+                downloadManager.addOfflinePlaylist(p);
+              }
+              Fluttertoast.showToast(
+                msg: 'Track added to'.i18n + ' ${p.title}',
+                toastLength: Toast.LENGTH_SHORT,
+                gravity: ToastGravity.BOTTOM,
+              );
+            },
+          );
         },
       );
+      if (context.mounted) _close(context);
+    },
+  );
 
   Widget removeFromPlaylist(Track t, Playlist p, BuildContext context) =>
       ListTile(
@@ -297,9 +276,10 @@ class MenuSheet {
           //Remove from cache
           cache.libraryTracks?.removeWhere((i) => i == t.id);
           Fluttertoast.showToast(
-              msg: 'Track removed from library'.i18n,
-              toastLength: Toast.LENGTH_SHORT,
-              gravity: ToastGravity.BOTTOM);
+            msg: 'Track removed from library'.i18n,
+            toastLength: Toast.LENGTH_SHORT,
+            gravity: ToastGravity.BOTTOM,
+          );
           if (onUpdate != null) onUpdate();
           if (context.mounted) _close(context);
         },
@@ -307,87 +287,93 @@ class MenuSheet {
 
   //Redirect to artist page (ie from track)
   Widget showArtist(Artist a, BuildContext context) => ListTile(
-        title: Text(
-          'Go to'.i18n + ' ${a.name}',
-          maxLines: 1,
-          overflow: TextOverflow.ellipsis,
-        ),
-        leading: const Icon(Icons.recent_actors),
-        onTap: () async {
-          try {
-            Artist b = await deezerAPI.artist(a.id!);
-            if (context.mounted) _close(context);
-            customNavigatorKey.currentState
-                ?.push(MaterialPageRoute(builder: (context) => ArtistDetails(b)));
+    title: Text(
+      'Go to'.i18n + ' ${a.name}',
+      maxLines: 1,
+      overflow: TextOverflow.ellipsis,
+    ),
+    leading: const Icon(Icons.recent_actors),
+    onTap: () async {
+      try {
+        Artist b = await deezerAPI.artist(a.id!);
+        if (context.mounted) _close(context);
+        customNavigatorKey.currentState?.push(
+          MaterialPageRoute(builder: (context) => ArtistDetails(b)),
+        );
 
-            navigateCallback();
-          } catch(e) {
-            if (context.mounted) _close(context);
-            customNavigatorKey.currentState
-                ?.push(MaterialPageRoute(builder: (context) => const ErrorScreen()));
+        navigateCallback();
+      } catch (e) {
+        if (context.mounted) _close(context);
+        customNavigatorKey.currentState?.push(
+          MaterialPageRoute(builder: (context) => const ErrorScreen()),
+        );
 
-            navigateCallback();
-          }
-        }
-      );
+        navigateCallback();
+      }
+    },
+  );
 
   Widget showAlbum(Album a, BuildContext context) => ListTile(
-        title: Text(
-          'Go to'.i18n + ' ${a.title}',
-          maxLines: 1,
-          overflow: TextOverflow.ellipsis,
-        ),
-        leading: const Icon(Icons.album),
-        onTap: () {
-          if (context.mounted) _close(context);
-          customNavigatorKey.currentState
-              ?.push(MaterialPageRoute(builder: (context) => AlbumDetails(a)));
-
-          navigateCallback();
-        },
+    title: Text(
+      'Go to'.i18n + ' ${a.title}',
+      maxLines: 1,
+      overflow: TextOverflow.ellipsis,
+    ),
+    leading: const Icon(Icons.album),
+    onTap: () {
+      if (context.mounted) _close(context);
+      customNavigatorKey.currentState?.push(
+        MaterialPageRoute(builder: (context) => AlbumDetails(a)),
       );
+
+      navigateCallback();
+    },
+  );
 
   Widget playMix(Track track, BuildContext context) => ListTile(
-        title: Text('Play mix'.i18n),
-        leading: const Icon(Icons.online_prediction),
+    title: Text('Play mix'.i18n),
+    leading: const Icon(Icons.online_prediction),
+    onTap: () async {
+      GetIt.I<AudioPlayerHandler>().playMix(track.id!, track.title!);
+      if (context.mounted) _close(context);
+    },
+  );
+
+  Widget offlineTrack(Track track, BuildContext context) => FutureBuilder(
+    future: downloadManager.checkOffline(track: track),
+    builder: (innerContext, snapshot) {
+      bool isOffline = snapshot.data ?? (track.offline ?? false);
+      return ListTile(
+        title: Text(isOffline ? 'Remove offline'.i18n : 'Offline'.i18n),
+        leading: const Icon(Icons.offline_pin),
         onTap: () async {
-          GetIt.I<AudioPlayerHandler>().playMix(track.id!, track.title!);
+          if (isOffline) {
+            await downloadManager.removeOfflineTracks([track]);
+            Fluttertoast.showToast(
+              msg: 'Track removed from offline!'.i18n,
+              gravity: ToastGravity.BOTTOM,
+              toastLength: Toast.LENGTH_SHORT,
+            );
+          } else {
+            await downloadManager.addOfflineTrack(track, private: true);
+          }
           if (context.mounted) _close(context);
         },
       );
-
-  Widget offlineTrack(Track track, BuildContext context) => FutureBuilder(
-        future: downloadManager.checkOffline(track: track),
-        builder: (innerContext, snapshot) {
-          bool isOffline = snapshot.data ?? (track.offline ?? false);
-          return ListTile(
-            title: Text(isOffline ? 'Remove offline'.i18n : 'Offline'.i18n),
-            leading: const Icon(Icons.offline_pin),
-            onTap: () async {
-              if (isOffline) {
-                await downloadManager.removeOfflineTracks([track]);
-                Fluttertoast.showToast(
-                    msg: 'Track removed from offline!'.i18n,
-                    gravity: ToastGravity.BOTTOM,
-                    toastLength: Toast.LENGTH_SHORT);
-              } else {
-                await downloadManager.addOfflineTrack(track, private: true);
-              }
-              if (context.mounted) _close(context);
-            },
-          );
-        },
-      );
+    },
+  );
 
   //===================
   // ALBUM
   //===================
 
   //Default album options
-  void defaultAlbumMenu(Album album,
-      {required BuildContext context,
-      List<Widget> options = const [],
-      Function? onRemove}) {
+  void defaultAlbumMenu(
+    Album album, {
+    required BuildContext context,
+    List<Widget> options = const [],
+    Function? onRemove,
+  }) {
     show(context, [
       (album.library != null && onRemove != null)
           ? removeAlbum(album, context, onRemove: onRemove)
@@ -395,7 +381,7 @@ class MenuSheet {
       downloadAlbum(album, context),
       offlineAlbum(album, context),
       shareTile('album', album.id!),
-      ...options
+      ...options,
     ]);
   }
 
@@ -404,70 +390,77 @@ class MenuSheet {
   //===================
 
   Widget downloadAlbum(Album a, BuildContext context) => ListTile(
-      title: Text('Download'.i18n),
-      leading: const Icon(Icons.file_download),
-      onTap: () async {
-        if (context.mounted) _close(context);
-        if (await downloadManager.addOfflineAlbum(a, private: false) != false) {
-          showDownloadStartedToast();
-        }
-      });
+    title: Text('Download'.i18n),
+    leading: const Icon(Icons.file_download),
+    onTap: () async {
+      if (context.mounted) _close(context);
+      if (await downloadManager.addOfflineAlbum(a, private: false) != false) {
+        showDownloadStartedToast();
+      }
+    },
+  );
 
   Widget offlineAlbum(Album a, BuildContext context) => ListTile(
-        title: Text('Make offline'.i18n),
-        leading: const Icon(Icons.offline_pin),
-        onTap: () async {
-          await deezerAPI.addFavoriteAlbum(a.id!);
-          await downloadManager.addOfflineAlbum(a, private: true);
-          if (context.mounted) _close(context);
-          showDownloadStartedToast();
-        },
-      );
+    title: Text('Make offline'.i18n),
+    leading: const Icon(Icons.offline_pin),
+    onTap: () async {
+      await deezerAPI.addFavoriteAlbum(a.id!);
+      await downloadManager.addOfflineAlbum(a, private: true);
+      if (context.mounted) _close(context);
+      showDownloadStartedToast();
+    },
+  );
 
   Widget libraryAlbum(Album a, BuildContext context) => ListTile(
-        title: Text('Add to library'.i18n),
-        leading: const Icon(Icons.library_music),
-        onTap: () async {
-          await deezerAPI.addFavoriteAlbum(a.id!);
-          Fluttertoast.showToast(
-              msg: 'Added to library'.i18n, gravity: ToastGravity.BOTTOM);
-          if (context.mounted) _close(context);
-        },
+    title: Text('Add to library'.i18n),
+    leading: const Icon(Icons.library_music),
+    onTap: () async {
+      await deezerAPI.addFavoriteAlbum(a.id!);
+      Fluttertoast.showToast(
+        msg: 'Added to library'.i18n,
+        gravity: ToastGravity.BOTTOM,
       );
+      if (context.mounted) _close(context);
+    },
+  );
 
   //Remove album from favorites
-  Widget removeAlbum(Album a, BuildContext context,
-          {required Function onRemove}) =>
-      ListTile(
-        title: Text('Remove album'.i18n),
-        leading: const Icon(Icons.delete),
-        onTap: () async {
-          await deezerAPI.removeAlbum(a.id!);
-          await downloadManager.removeOfflineAlbum(a.id!);
-          Fluttertoast.showToast(
-            msg: 'Album removed'.i18n,
-            toastLength: Toast.LENGTH_SHORT,
-            gravity: ToastGravity.BOTTOM,
-          );
-          onRemove();
-          if (context.mounted) _close(context);
-        },
+  Widget removeAlbum(
+    Album a,
+    BuildContext context, {
+    required Function onRemove,
+  }) => ListTile(
+    title: Text('Remove album'.i18n),
+    leading: const Icon(Icons.delete),
+    onTap: () async {
+      await deezerAPI.removeAlbum(a.id!);
+      await downloadManager.removeOfflineAlbum(a.id!);
+      Fluttertoast.showToast(
+        msg: 'Album removed'.i18n,
+        toastLength: Toast.LENGTH_SHORT,
+        gravity: ToastGravity.BOTTOM,
       );
+      onRemove();
+      if (context.mounted) _close(context);
+    },
+  );
 
   //===================
   // ARTIST
   //===================
 
-  void defaultArtistMenu(Artist artist,
-      {required BuildContext context,
-      List<Widget> options = const [],
-      Function? onRemove}) {
+  void defaultArtistMenu(
+    Artist artist, {
+    required BuildContext context,
+    List<Widget> options = const [],
+    Function? onRemove,
+  }) {
     show(context, [
       (artist.library != null)
           ? removeArtist(artist, context, onRemove: onRemove)
           : favoriteArtist(artist, context),
       shareTile('artist', artist.id!),
-      ...options
+      ...options,
     ]);
   }
 
@@ -482,36 +475,40 @@ class MenuSheet {
         onTap: () async {
           await deezerAPI.removeArtist(a.id!);
           Fluttertoast.showToast(
-              msg: 'Artist removed from library'.i18n,
-              toastLength: Toast.LENGTH_SHORT,
-              gravity: ToastGravity.BOTTOM);
+            msg: 'Artist removed from library'.i18n,
+            toastLength: Toast.LENGTH_SHORT,
+            gravity: ToastGravity.BOTTOM,
+          );
           if (onRemove != null) onRemove();
           if (context.mounted) _close(context);
         },
       );
 
   Widget favoriteArtist(Artist a, BuildContext context) => ListTile(
-        title: Text('Add to favorites'.i18n),
-        leading: const Icon(Icons.favorite),
-        onTap: () async {
-          await deezerAPI.addFavoriteArtist(a.id!);
-          Fluttertoast.showToast(
-              msg: 'Added to library'.i18n,
-              toastLength: Toast.LENGTH_SHORT,
-              gravity: ToastGravity.BOTTOM);
-          if (context.mounted) _close(context);
-        },
+    title: Text('Add to favorites'.i18n),
+    leading: const Icon(Icons.favorite),
+    onTap: () async {
+      await deezerAPI.addFavoriteArtist(a.id!);
+      Fluttertoast.showToast(
+        msg: 'Added to library'.i18n,
+        toastLength: Toast.LENGTH_SHORT,
+        gravity: ToastGravity.BOTTOM,
       );
+      if (context.mounted) _close(context);
+    },
+  );
 
   //===================
   // PLAYLIST
   //===================
 
-  void defaultPlaylistMenu(Playlist playlist,
-      {required BuildContext context,
-      List<Widget> options = const [],
-      Function? onRemove,
-      Function? onUpdate}) {
+  void defaultPlaylistMenu(
+    Playlist playlist, {
+    required BuildContext context,
+    List<Widget> options = const [],
+    Function? onRemove,
+    Function? onUpdate,
+  }) {
     show(context, [
       (playlist.library != null)
           ? removePlaylistLibrary(playlist, context, onRemove: onRemove)
@@ -521,7 +518,7 @@ class MenuSheet {
       shareTile('playlist', playlist.id!),
       if (playlist.user?.id == deezerAPI.userId)
         editPlaylist(playlist, context: context, onUpdate: onUpdate),
-      ...options
+      ...options,
     ]);
   }
 
@@ -529,100 +526,117 @@ class MenuSheet {
   // PLAYLIST OPTIONS
   //===================
 
-  Widget removePlaylistLibrary(Playlist p, BuildContext context,
-          {Function? onRemove}) =>
-      ListTile(
-        title: Text('Remove from library'.i18n),
-        leading: const Icon(Icons.delete),
-        onTap: () async {
-          if (p.user?.id?.trim() == deezerAPI.userId) {
-            //Delete playlist if own
-            await deezerAPI.deletePlaylist(p.id!);
-          } else {
-            //Just remove from library
-            await deezerAPI.removePlaylist(p.id!);
-          }
-          downloadManager.removeOfflinePlaylist(p.id!);
-          if (onRemove != null) onRemove();
-          if (context.mounted) _close(context);
-        },
-      );
+  Widget removePlaylistLibrary(
+    Playlist p,
+    BuildContext context, {
+    Function? onRemove,
+  }) => ListTile(
+    title: Text('Remove from library'.i18n),
+    leading: const Icon(Icons.delete),
+    onTap: () async {
+      if (p.user?.id?.trim() == deezerAPI.userId) {
+        //Delete playlist if own
+        await deezerAPI.deletePlaylist(p.id!);
+      } else {
+        //Just remove from library
+        await deezerAPI.removePlaylist(p.id!);
+      }
+      downloadManager.removeOfflinePlaylist(p.id!);
+      if (onRemove != null) onRemove();
+      if (context.mounted) _close(context);
+    },
+  );
 
   Widget addPlaylistLibrary(Playlist p, BuildContext context) => ListTile(
-        title: Text('Add playlist to library'.i18n),
-        leading: const Icon(Icons.favorite),
-        onTap: () async {
-          await deezerAPI.addPlaylist(p.id!);
-          Fluttertoast.showToast(
-              msg: 'Added playlist to library'.i18n,
-              gravity: ToastGravity.BOTTOM);
-          if (context.mounted) _close(context);
-        },
+    title: Text('Add playlist to library'.i18n),
+    leading: const Icon(Icons.favorite),
+    onTap: () async {
+      await deezerAPI.addPlaylist(p.id!);
+      Fluttertoast.showToast(
+        msg: 'Added playlist to library'.i18n,
+        gravity: ToastGravity.BOTTOM,
       );
+      if (context.mounted) _close(context);
+    },
+  );
 
   Widget addPlaylistOffline(Playlist p, BuildContext context) => ListTile(
-        title: Text('Make playlist offline'.i18n),
-        leading: const Icon(Icons.offline_pin),
-        onTap: () async {
-          //Add to library
-          await deezerAPI.addPlaylist(p.id!);
-          downloadManager.addOfflinePlaylist(p, private: true);
-          if (context.mounted) _close(context);
-          showDownloadStartedToast();
-        },
-      );
+    title: Text('Make playlist offline'.i18n),
+    leading: const Icon(Icons.offline_pin),
+    onTap: () async {
+      //Add to library
+      await deezerAPI.addPlaylist(p.id!);
+      downloadManager.addOfflinePlaylist(p, private: true);
+      if (context.mounted) _close(context);
+      showDownloadStartedToast();
+    },
+  );
 
   Widget downloadPlaylist(Playlist p, BuildContext context) => ListTile(
-        title: Text('Download playlist'.i18n),
-        leading: const Icon(Icons.file_download),
-        onTap: () async {
-          if (context.mounted) _close(context);
-          if (await downloadManager.addOfflinePlaylist(p, private: false) !=
-              false) {
-            showDownloadStartedToast();
-          }
-        },
-      );
+    title: Text('Download playlist'.i18n),
+    leading: const Icon(Icons.file_download),
+    onTap: () async {
+      if (context.mounted) _close(context);
+      if (await downloadManager.addOfflinePlaylist(p, private: false) !=
+          false) {
+        showDownloadStartedToast();
+      }
+    },
+  );
 
-  Widget editPlaylist(Playlist p,
-          {required BuildContext context, Function? onUpdate}) =>
-      ListTile(
-        title: Text('Edit playlist'.i18n),
-        leading: const Icon(Icons.edit),
-        onTap: () async {
-          await showDialog(
-              context: context,
-              builder: (context) => CreatePlaylistDialog(playlist: p));
-          if (context.mounted) _close(context);
-          if (onUpdate != null) onUpdate();
-        },
+  Widget editPlaylist(
+    Playlist p, {
+    required BuildContext context,
+    Function? onUpdate,
+  }) => ListTile(
+    title: Text('Edit playlist'.i18n),
+    leading: const Icon(Icons.edit),
+    onTap: () async {
+      await showDialog(
+        context: context,
+        builder: (context) => CreatePlaylistDialog(playlist: p),
       );
+      if (context.mounted) _close(context);
+      if (onUpdate != null) onUpdate();
+    },
+  );
 
   //===================
   // SHOW/EPISODE
   //===================
 
-  defaultShowMenu(Show s,
-      {required BuildContext context, List<Widget> options = const [], Function? onRemove, Function? onUpdate}) async {
+  defaultShowMenu(
+    Show s, {
+    required BuildContext context,
+    List<Widget> options = const [],
+    Function? onRemove,
+    Function? onUpdate,
+  }) async {
     show(context, [
       (await deezerAPI.checkShowFavorite(s))
           ? removeShowLibrary(s, context, onRemove: onRemove)
           : addShowLibrary(s, context),
       shareShow(s.id!),
-      ...options
+      ...options,
     ]);
   }
 
-  defaultShowEpisodeMenu(Show s, ShowEpisode e,
-      {required BuildContext context, List<Widget> options = const [], Function? onRemove, Function? onUpdate}) async {
+  defaultShowEpisodeMenu(
+    Show s,
+    ShowEpisode e, {
+    required BuildContext context,
+    List<Widget> options = const [],
+    Function? onRemove,
+    Function? onUpdate,
+  }) async {
     show(context, [
       (await deezerAPI.checkShowFavorite(s))
-      ? removeShowLibrary(s, context, onRemove: onRemove)
-      : addShowLibrary(s, context),
+          ? removeShowLibrary(s, context, onRemove: onRemove)
+          : addShowLibrary(s, context),
       shareTile('episode', e.id!),
       shareShow(s.id!),
       downloadExternalEpisode(e),
-      ...options
+      ...options,
     ]);
   }
 
@@ -631,45 +645,50 @@ class MenuSheet {
   //===================
 
   Widget shareShow(String id) => ListTile(
-        title: Text('Share show'.i18n),
-        leading: const Icon(Icons.share),
-        onTap: () async {
-          Share.share('https://deezer.com/show/$id');
-        },
+    title: Text('Share show'.i18n),
+    leading: const Icon(Icons.share),
+    onTap: () async {
+      SharePlus.instance.share(
+        ShareParams(uri: Uri.parse('https://deezer.com/show/$id')),
       );
+    },
+  );
 
-  Widget removeShowLibrary(Show s, BuildContext context,
-          {Function? onRemove}) =>
-      ListTile(
-        title: Text('Remove from library'.i18n),
-        leading: const Icon(Icons.delete),
-        onTap: () async {
-          await deezerAPI.removeShow(s.id!);
-          if (onRemove != null) onRemove();
-          if (context.mounted) _close(context);
-        },
-      );
+  Widget removeShowLibrary(
+    Show s,
+    BuildContext context, {
+    Function? onRemove,
+  }) => ListTile(
+    title: Text('Remove from library'.i18n),
+    leading: const Icon(Icons.delete),
+    onTap: () async {
+      await deezerAPI.removeShow(s.id!);
+      if (onRemove != null) onRemove();
+      if (context.mounted) _close(context);
+    },
+  );
 
   Widget addShowLibrary(Show s, BuildContext context) => ListTile(
-        title: Text('Add podcast to library'.i18n),
-        leading: const Icon(Icons.favorite),
-        onTap: () async {
-          await deezerAPI.addFavoriteShow(s.id!);
-          Fluttertoast.showToast(
-              msg: 'Added Show to library'.i18n,
-              gravity: ToastGravity.BOTTOM);
-          if (context.mounted) _close(context);
-        },
+    title: Text('Add podcast to library'.i18n),
+    leading: const Icon(Icons.favorite),
+    onTap: () async {
+      await deezerAPI.addFavoriteShow(s.id!);
+      Fluttertoast.showToast(
+        msg: 'Added Show to library'.i18n,
+        gravity: ToastGravity.BOTTOM,
       );
+      if (context.mounted) _close(context);
+    },
+  );
 
   //Open direct download link in browser
   Widget downloadExternalEpisode(ShowEpisode e) => ListTile(
-        title: Text('Download externally'.i18n),
-        leading: const Icon(Icons.file_download),
-        onTap: () async {
-          if (e.url != null) await launchUrlString(e.url!);
-        },
-      );
+    title: Text('Download externally'.i18n),
+    leading: const Icon(Icons.file_download),
+    onTap: () async {
+      if (e.url != null) await launchUrlString(e.url!);
+    },
+  );
 
   //===================
   // OTHER
@@ -677,60 +696,69 @@ class MenuSheet {
 
   showDownloadStartedToast() {
     Fluttertoast.showToast(
-        msg: 'Downloads added!'.i18n,
-        gravity: ToastGravity.BOTTOM,
-        toastLength: Toast.LENGTH_SHORT);
+      msg: 'Downloads added!'.i18n,
+      gravity: ToastGravity.BOTTOM,
+      toastLength: Toast.LENGTH_SHORT,
+    );
   }
 
   //Create playlist
   Future createPlaylist(BuildContext context) async {
     await showDialog(
-        context: context,
-        builder: (BuildContext context) {
-          return const CreatePlaylistDialog();
-        });
+      context: context,
+      builder: (BuildContext context) {
+        return const CreatePlaylistDialog();
+      },
+    );
   }
 
   Widget shareTile(String type, String id) => ListTile(
-        title: Text('Share'.i18n),
-        leading: const Icon(Icons.share),
-        onTap: () async {
-          Share.share('https://deezer.com/$type/$id');
-        },
+    title: Text('Share'.i18n),
+    leading: const Icon(Icons.share),
+    onTap: () async {
+      SharePlus.instance.share(
+        ShareParams(uri: Uri.parse('https://deezer.com/$type/$id')),
       );
+    },
+  );
 
   Widget sleepTimer(BuildContext context) => ListTile(
-        title: Text('Sleep timer'.i18n),
-        leading: const Icon(Icons.access_time),
-        onTap: () async {
-          showDialog(
-              context: context,
-              builder: (context) {
-                return const SleepTimerDialog();
-              });
+    title: Text('Sleep timer'.i18n),
+    leading: const Icon(Icons.access_time),
+    onTap: () async {
+      showDialog(
+        context: context,
+        builder: (context) {
+          return const SleepTimerDialog();
         },
       );
+    },
+  );
 
   Widget wakelock(BuildContext context) => ListTile(
-        title: Text('Keep the screen on'.i18n),
-        leading: const Icon(Icons.screen_lock_portrait),
-        onTap: () async {
-          _close(context);
-          //Enable
-          if (!cache.wakelock) {
-            WakelockPlus.enable();
-            Fluttertoast.showToast(
-                msg: 'Wakelock enabled!'.i18n, gravity: ToastGravity.BOTTOM);
-            cache.wakelock = true;
-            return;
-          }
-          //Disable
-          WakelockPlus.disable();
-          Fluttertoast.showToast(
-              msg: 'Wakelock disabled!'.i18n, gravity: ToastGravity.BOTTOM);
-          cache.wakelock = false;
-        },
+    title: Text('Keep the screen on'.i18n),
+    leading: const Icon(Icons.screen_lock_portrait),
+    onTap: () async {
+      _close(context);
+      //Enable
+      if (!cache.wakelock) {
+        WakelockPlus.enable();
+        Fluttertoast.showToast(
+          msg: 'Wakelock enabled!'.i18n,
+          gravity: ToastGravity.BOTTOM,
+        );
+        cache.wakelock = true;
+        return;
+      }
+      //Disable
+      WakelockPlus.disable();
+      Fluttertoast.showToast(
+        msg: 'Wakelock disabled!'.i18n,
+        gravity: ToastGravity.BOTTOM,
       );
+      cache.wakelock = false;
+    },
+  );
 
   void _close(BuildContext context) => Navigator.of(context).pop();
 }
@@ -765,10 +793,11 @@ class _SleepTimerDialogState extends State<SleepTimerDialog> {
                 children: [
                   Text('Hours:'.i18n),
                   NumberPicker(
-                      value: hours,
-                      minValue: 0,
-                      maxValue: 69,
-                      onChanged: (v) => setState(() => hours = v)),
+                    value: hours,
+                    minValue: 0,
+                    maxValue: 69,
+                    onChanged: (v) => setState(() => hours = v),
+                  ),
                 ],
               ),
               Column(
@@ -776,10 +805,11 @@ class _SleepTimerDialogState extends State<SleepTimerDialog> {
                 children: [
                   Text('Minutes:'.i18n),
                   NumberPicker(
-                      value: minutes,
-                      minValue: 0,
-                      maxValue: 60,
-                      onChanged: (v) => setState(() => minutes = v)),
+                    value: minutes,
+                    minValue: 0,
+                    maxValue: 60,
+                    onChanged: (v) => setState(() => minutes = v),
+                  ),
                 ],
               ),
             ],
@@ -789,14 +819,21 @@ class _SleepTimerDialogState extends State<SleepTimerDialog> {
             Text(
               'Current timer ends at'.i18n + ': ' + _endTime(),
               textAlign: TextAlign.center,
-            )
+            ),
         ],
       ),
       actions: [
         TextButton(
-                  style: ButtonStyle(
-          overlayColor: WidgetStateProperty.resolveWith<Color?>((Set<WidgetState> states) {if (states.contains(WidgetState.pressed)) {return Theme.of(context).primaryColor.withOpacity(0.3);}return null;}),
-         ),
+          style: ButtonStyle(
+            overlayColor: WidgetStateProperty.resolveWith<Color?>((
+              Set<WidgetState> states,
+            ) {
+              if (states.contains(WidgetState.pressed)) {
+                return Theme.of(context).primaryColor.withValues(alpha: 0.3);
+              }
+              return null;
+            }),
+          ),
           child: Text('Dismiss'.i18n),
           onPressed: () {
             Navigator.of(context).pop();
@@ -804,9 +841,16 @@ class _SleepTimerDialogState extends State<SleepTimerDialog> {
         ),
         if (cache.sleepTimer != null)
           TextButton(
-                    style: ButtonStyle(
-          overlayColor: WidgetStateProperty.resolveWith<Color?>((Set<WidgetState> states) {if (states.contains(WidgetState.pressed)) {return Theme.of(context).primaryColor.withOpacity(0.3);}return null;}),
-         ),
+            style: ButtonStyle(
+              overlayColor: WidgetStateProperty.resolveWith<Color?>((
+                Set<WidgetState> states,
+              ) {
+                if (states.contains(WidgetState.pressed)) {
+                  return Theme.of(context).primaryColor.withValues(alpha: 0.3);
+                }
+                return null;
+              }),
+            ),
             child: Text('Cancel current timer'.i18n),
             onPressed: () {
               cache.sleepTimer!.cancel();
@@ -816,21 +860,28 @@ class _SleepTimerDialogState extends State<SleepTimerDialog> {
             },
           ),
         TextButton(
-                  style: ButtonStyle(
-          overlayColor: WidgetStateProperty.resolveWith<Color?>((Set<WidgetState> states) {if (states.contains(WidgetState.pressed)) {return Theme.of(context).primaryColor.withOpacity(0.3);}return null;}),
-         ),
+          style: ButtonStyle(
+            overlayColor: WidgetStateProperty.resolveWith<Color?>((
+              Set<WidgetState> states,
+            ) {
+              if (states.contains(WidgetState.pressed)) {
+                return Theme.of(context).primaryColor.withValues(alpha: 0.3);
+              }
+              return null;
+            }),
+          ),
           child: Text('Save'.i18n),
           onPressed: () {
             Duration duration = Duration(hours: hours, minutes: minutes);
             cache.sleepTimer?.cancel();
             //Create timer
-            cache.sleepTimer =
-                Stream.fromFuture(Future.delayed(duration)).listen((_) {
-              GetIt.I<AudioPlayerHandler>().pause();
-              cache.sleepTimer?.cancel();
-              cache.sleepTimerTime = null;
-              cache.sleepTimer = null;
-            });
+            cache.sleepTimer = Stream.fromFuture(Future.delayed(duration))
+                .listen((_) {
+                  GetIt.I<AudioPlayerHandler>().pause();
+                  cache.sleepTimer?.cancel();
+                  cache.sleepTimerTime = null;
+                  cache.sleepTimer = null;
+                });
             cache.sleepTimerTime = DateTime.now().add(duration);
             Navigator.of(context).pop();
           },
@@ -868,52 +919,51 @@ class _SelectPlaylistDialogState extends State<SelectPlaylistDialog> {
         future: deezerAPI.getPlaylists(),
         builder: (context, snapshot) {
           if (snapshot.hasError) {
-            const SizedBox(
-              height: 100,
-              child: ErrorScreen(),
-            );
+            const SizedBox(height: 100, child: ErrorScreen());
           }
           if (snapshot.connectionState != ConnectionState.done) {
             return SizedBox(
               height: 100,
               child: Center(
-                child: CircularProgressIndicator(color: Theme.of(context).primaryColor,),
+                child: CircularProgressIndicator(
+                  color: Theme.of(context).primaryColor,
+                ),
               ),
             );
           }
 
           // Check if snapshot has data before accessing it
           if (!snapshot.hasData || snapshot.data!.isEmpty) {
-            return const SizedBox(
-              child: ErrorScreen(),
-            );  
+            return const SizedBox(child: ErrorScreen());
           }
 
           List<Playlist> playlists = snapshot.data!;
           return SingleChildScrollView(
-            child: Column(mainAxisSize: MainAxisSize.min, children: [
-              ...List.generate(
+            child: Column(
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                ...List.generate(
                   playlists.length,
                   (i) => ListTile(
-                        title: Text(playlists[i].title!),
-                        leading: CachedImage(
-                          url: playlists[i].image?.thumb ?? '',
-                        ),
-                        onTap: () {
-                          widget.callback(playlists[i]);
-                          Navigator.of(context).pop();
-                        },
-                      )),
-              ListTile(
-                title: Text('Create new playlist'.i18n),
-                leading: const Icon(Icons.add),
-                onTap: () async {
-                  setState(() {
-                    createNew = true;
-                  });
-                },
-              )
-            ]),
+                    title: Text(playlists[i].title!),
+                    leading: CachedImage(url: playlists[i].image?.thumb ?? ''),
+                    onTap: () {
+                      widget.callback(playlists[i]);
+                      Navigator.of(context).pop();
+                    },
+                  ),
+                ),
+                ListTile(
+                  title: Text('Create new playlist'.i18n),
+                  leading: const Icon(Icons.add),
+                  onTap: () async {
+                    setState(() {
+                      createNew = true;
+                    });
+                  },
+                ),
+              ],
+            ),
           );
         },
       ),
@@ -964,10 +1014,13 @@ class _CreatePlaylistDialogState extends State<CreatePlaylistDialog> {
         children: <Widget>[
           TextField(
             cursorColor: Theme.of(context).primaryColor,
-            decoration: InputDecoration(labelText: 'Title'.i18n,
+            decoration: InputDecoration(
+              labelText: 'Title'.i18n,
               focusedBorder: UnderlineInputBorder(
-              borderSide: BorderSide(color: Theme.of(context).primaryColor), // Color of the underline when focused
-            ),
+                borderSide: BorderSide(
+                  color: Theme.of(context).primaryColor,
+                ), // Color of the underline when focused
+              ),
             ),
             controller: _titleController ?? TextEditingController(),
             onChanged: (String s) => _title = s,
@@ -976,25 +1029,23 @@ class _CreatePlaylistDialogState extends State<CreatePlaylistDialog> {
             cursorColor: Theme.of(context).primaryColor,
             onChanged: (String s) => _description = s,
             controller: _descController ?? TextEditingController(),
-            decoration: InputDecoration(labelText: 'Description'.i18n,
-            focusedBorder: UnderlineInputBorder(
-              borderSide: BorderSide(color: Theme.of(context).primaryColor), // Color of the underline when focused
-            ),
+            decoration: InputDecoration(
+              labelText: 'Description'.i18n,
+              focusedBorder: UnderlineInputBorder(
+                borderSide: BorderSide(
+                  color: Theme.of(context).primaryColor,
+                ), // Color of the underline when focused
+              ),
             ),
           ),
-          Container(
-            height: 4.0,
-          ),
+          Container(height: 4.0),
           DropdownButton<int>(
             value: _playlistType,
             onChanged: (int? v) {
               setState(() => _playlistType = v!);
             },
             items: [
-              DropdownMenuItem<int>(
-                value: 1,
-                child: Text('Private'.i18n),
-              ),
+              DropdownMenuItem<int>(value: 1, child: Text('Private'.i18n)),
               DropdownMenuItem<int>(
                 value: 2,
                 child: Text('Collaborative'.i18n),
@@ -1005,38 +1056,61 @@ class _CreatePlaylistDialogState extends State<CreatePlaylistDialog> {
       ),
       actions: <Widget>[
         TextButton(
-                  style: ButtonStyle(
-          overlayColor: WidgetStateProperty.resolveWith<Color?>((Set<WidgetState> states) {if (states.contains(WidgetState.pressed)) {return Theme.of(context).primaryColor.withOpacity(0.3);}return null;}),
-         ),
+          style: ButtonStyle(
+            overlayColor: WidgetStateProperty.resolveWith<Color?>((
+              Set<WidgetState> states,
+            ) {
+              if (states.contains(WidgetState.pressed)) {
+                return Theme.of(context).primaryColor.withValues(alpha: 0.3);
+              }
+              return null;
+            }),
+          ),
           child: Text('Cancel'.i18n),
           onPressed: () => Navigator.of(context).pop(),
         ),
         TextButton(
-                  style: ButtonStyle(
-          overlayColor: WidgetStateProperty.resolveWith<Color?>((Set<WidgetState> states) {if (states.contains(WidgetState.pressed)) {return Theme.of(context).primaryColor.withOpacity(0.3);}return null;}),
-         ),
+          style: ButtonStyle(
+            overlayColor: WidgetStateProperty.resolveWith<Color?>((
+              Set<WidgetState> states,
+            ) {
+              if (states.contains(WidgetState.pressed)) {
+                return Theme.of(context).primaryColor.withValues(alpha: 0.3);
+              }
+              return null;
+            }),
+          ),
           child: Text(edit ? 'Update'.i18n : 'Create'.i18n),
           onPressed: () async {
             if (edit) {
               //Update
-              await deezerAPI.updatePlaylist(widget.playlist!.id!,
-                  _titleController!.value.text, _descController!.value.text,
-                  status: _playlistType);
+              await deezerAPI.updatePlaylist(
+                widget.playlist!.id!,
+                _titleController!.value.text,
+                _descController!.value.text,
+                status: _playlistType,
+              );
               Fluttertoast.showToast(
-                  msg: 'Playlist updated!'.i18n, gravity: ToastGravity.BOTTOM);
+                msg: 'Playlist updated!'.i18n,
+                gravity: ToastGravity.BOTTOM,
+              );
             } else {
               List<String> tracks = [];
               tracks = widget.tracks?.map<String>((t) => t.id!).toList() ?? [];
-              await deezerAPI.createPlaylist(_title,
-                  status: _playlistType,
-                  description: _description,
-                  trackIds: tracks);
+              await deezerAPI.createPlaylist(
+                _title,
+                status: _playlistType,
+                description: _description,
+                trackIds: tracks,
+              );
               Fluttertoast.showToast(
-                  msg: 'Playlist created!'.i18n, gravity: ToastGravity.BOTTOM);
+                msg: 'Playlist created!'.i18n,
+                gravity: ToastGravity.BOTTOM,
+              );
             }
             if (context.mounted) Navigator.of(context).pop();
           },
-        )
+        ),
       ],
     );
   }
