@@ -8,6 +8,7 @@ import 'package:disk_space_plus/disk_space_plus.dart';
 import 'package:external_path/external_path.dart';
 import 'package:filesize/filesize.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/material.dart' as material;
 import 'package:flutter_displaymode/flutter_displaymode.dart';
 //import 'package:flutter_file_dialog/flutter_file_dialog.dart';
 import 'package:flutter_material_color_picker/flutter_material_color_picker.dart';
@@ -55,33 +56,50 @@ class _SettingsScreenState extends State<SettingsScreen> {
         children: <Widget>[
           ListTile(
             title: Text('General'.i18n),
-            leading: const LeadingIcon(Icons.settings, color: Color(0xffeca704)),
-            onTap: () => Navigator.of(context).push(MaterialPageRoute(builder: (context) => const GeneralSettings())),
+            leading:
+                const LeadingIcon(Icons.settings, color: Color(0xffeca704)),
+            onTap: () => Navigator.of(context).push(MaterialPageRoute(
+                builder: (context) => const GeneralSettings())),
           ),
           ListTile(
             title: Text('Download Settings'.i18n),
-            leading: const LeadingIcon(Icons.cloud_download, color: Color(0xffbe3266)),
-            onTap: () => Navigator.of(context).push(MaterialPageRoute(builder: (context) => const DownloadsSettings())),
+            leading: const LeadingIcon(Icons.cloud_download,
+                color: Color(0xffbe3266)),
+            onTap: () => Navigator.of(context).push(MaterialPageRoute(
+                builder: (context) => const DownloadsSettings())),
           ),
           ListTile(
             title: Text('Appearance'.i18n),
-            leading: const LeadingIcon(Icons.color_lens, color: Color(0xff4b2e7e)),
-            onTap: () => Navigator.push(context, MaterialPageRoute(builder: (context) => const AppearanceSettings())),
+            leading:
+                const LeadingIcon(Icons.color_lens, color: Color(0xff4b2e7e)),
+            onTap: () => Navigator.push(
+                context,
+                MaterialPageRoute(
+                    builder: (context) => const AppearanceSettings())),
           ),
           ListTile(
             title: Text('Quality'.i18n),
-            leading: const LeadingIcon(Icons.high_quality, color: Color(0xff384697)),
-            onTap: () => Navigator.push(context, MaterialPageRoute(builder: (context) => const QualitySettings())),
+            leading:
+                const LeadingIcon(Icons.high_quality, color: Color(0xff384697)),
+            onTap: () => Navigator.push(
+                context,
+                MaterialPageRoute(
+                    builder: (context) => const QualitySettings())),
           ),
           ListTile(
             title: Text('Deezer'.i18n),
-            leading: const LeadingIcon(Icons.equalizer, color: Color(0xff0880b5)),
-            onTap: () => Navigator.push(context, MaterialPageRoute(builder: (context) => const DeezerSettings())),
+            leading:
+                const LeadingIcon(Icons.equalizer, color: Color(0xff0880b5)),
+            onTap: () => Navigator.push(
+                context,
+                MaterialPageRoute(
+                    builder: (context) => const DeezerSettings())),
           ),
           //Language select
           ListTile(
             title: Text('Language'.i18n),
-            leading: const LeadingIcon(Icons.language, color: Color(0xff009a85)),
+            leading:
+                const LeadingIcon(Icons.language, color: Color(0xff009a85)),
             onTap: () {
               showDialog(
                   context: context,
@@ -93,8 +111,10 @@ class _SettingsScreenState extends State<SettingsScreen> {
                           title: Text(l.name),
                           subtitle: Text('${l.locale}-${l.country}'),
                           onTap: () async {
-                            I18n.of(customNavigatorKey.currentContext!).locale = Locale(l.locale, l.country);
-                            setState(() => settings.language = '${l.locale}_${l.country}');
+                            I18n.of(customNavigatorKey.currentContext!).locale =
+                                Locale(l.locale, l.country);
+                            setState(() =>
+                                settings.language = '${l.locale}_${l.country}');
                             await settings.save();
                             // Close the SimpleDialog
                             if (context.mounted) Navigator.of(context).pop();
@@ -106,12 +126,14 @@ class _SettingsScreenState extends State<SettingsScreen> {
           ListTile(
             title: Text('Updates'.i18n),
             leading: const LeadingIcon(Icons.update, color: Color(0xff2ba766)),
-            onTap: () => Navigator.push(context, MaterialPageRoute(builder: (context) => UpdaterScreen())),
+            onTap: () => Navigator.push(context,
+                MaterialPageRoute(builder: (context) => const UpdaterScreen())),
           ),
           ListTile(
             title: Text('About'.i18n),
             leading: const LeadingIcon(Icons.info, color: Colors.grey),
-            onTap: () => Navigator.push(context, MaterialPageRoute(builder: (context) => const CreditsScreen())),
+            onTap: () => Navigator.push(context,
+                MaterialPageRoute(builder: (context) => const CreditsScreen())),
           ),
         ],
       ),
@@ -137,7 +159,8 @@ class _AppearanceSettingsState extends State<AppearanceSettings> {
         children: <Widget>[
           ListTile(
             title: Text('Theme'.i18n),
-            subtitle: Text('Currently'.i18n + ': ${settings.theme.toString().split('.').last}'),
+            subtitle: Text('Currently'.i18n +
+                ': ${settings.theme.toString().split('.').last}'),
             leading: const Icon(Icons.color_lens),
             onTap: () {
               showDialog(
@@ -205,7 +228,10 @@ class _AppearanceSettingsState extends State<AppearanceSettings> {
             leading: const Icon(Icons.font_download),
             subtitle: Text(settings.font),
             onTap: () {
-              showDialog(context: context, builder: (context) => FontSelector(() => Navigator.of(context).pop()));
+              showDialog(
+                  context: context,
+                  builder: (context) =>
+                      FontSelector(() => Navigator.of(context).pop()));
             },
           ),
           ListTile(
@@ -235,7 +261,8 @@ class _AppearanceSettingsState extends State<AppearanceSettings> {
           ),
           ListTile(
             title: Text('Theme Additonal Items'.i18n),
-            subtitle: Text('Themes additional items like mini player and lyrics'.i18n),
+            subtitle: Text(
+                'Themes additional items like mini player and lyrics'.i18n),
             leading: const Icon(Icons.library_add),
             trailing: Switch(
               value: settings.themeAdditonalItems,
@@ -248,7 +275,9 @@ class _AppearanceSettingsState extends State<AppearanceSettings> {
           ),
           ListTile(
             title: Text('Visualizer'.i18n),
-            subtitle: Text('Show visualizers on lyrics page. WARNING: Requires microphone permission!'.i18n),
+            subtitle: Text(
+                'Show visualizers on lyrics page. WARNING: Requires microphone permission!'
+                    .i18n),
             leading: const Icon(Icons.equalizer),
             trailing: Switch(
               value: settings.lyricsVisualizer,
@@ -335,7 +364,8 @@ class _AppearanceSettingsState extends State<AppearanceSettings> {
                                   onPressed: () async {
                                     settings.displayMode = i;
                                     await settings.save();
-                                    await FlutterDisplayMode.setPreferredMode(modes[i]);
+                                    await FlutterDisplayMode.setPreferredMode(
+                                        modes[i]);
                                     if (context.mounted) {
                                       Navigator.of(context).pop();
                                     }
@@ -362,7 +392,9 @@ class FontSelector extends StatefulWidget {
 class _FontSelectorState extends State<FontSelector> {
   String query = '';
   List<String> get fonts {
-    return settings.fonts.where((f) => f.toLowerCase().contains(query)).toList();
+    return settings.fonts
+        .where((f) => f.toLowerCase().contains(query))
+        .toList();
   }
 
   //Font selected
@@ -376,9 +408,18 @@ class _FontSelectorState extends State<FontSelector> {
                       .i18n),
               actions: [
                 TextButton(
-                          style: ButtonStyle(
-          overlayColor: WidgetStateProperty.resolveWith<Color?>((Set<WidgetState> states) {if (states.contains(WidgetState.pressed)) {return Theme.of(context).primaryColor.withOpacity(0.3);}return null;}),
-         ),
+                  style: ButtonStyle(
+                    overlayColor:
+                        material.WidgetStateProperty.resolveWith<Color?>(
+                            (Set<material.WidgetState> states) {
+                      if (states.contains(material.WidgetState.pressed)) {
+                        return Theme.of(context)
+                            .primaryColor
+                            .withValues(alpha: 0.3);
+                      }
+                      return null;
+                    }),
+                  ),
                   onPressed: () async {
                     setState(() => settings.font = font);
                     await settings.save();
@@ -390,9 +431,7 @@ class _FontSelectorState extends State<FontSelector> {
                   child: Text('Apply'.i18n),
                 ),
                 TextButton(
-                          style: ButtonStyle(
-          
-         ),
+                  style: const ButtonStyle(),
                   onPressed: () {
                     Navigator.of(context).pop();
                     widget.callback();
@@ -443,25 +482,29 @@ class _QualitySettingsState extends State<QualitySettings> {
         children: <Widget>[
           ListTile(
             title: Text('Mobile streaming'.i18n),
-            leading: const LeadingIcon(Icons.network_cell, color: Color(0xff384697)),
+            leading:
+                const LeadingIcon(Icons.network_cell, color: Color(0xff384697)),
           ),
           const QualityPicker('mobile'),
           const FreezerDivider(),
           ListTile(
             title: Text('Wifi streaming'.i18n),
-            leading: const LeadingIcon(Icons.network_wifi, color: Color(0xff0880b5)),
+            leading:
+                const LeadingIcon(Icons.network_wifi, color: Color(0xff0880b5)),
           ),
           const QualityPicker('wifi'),
           const FreezerDivider(),
           ListTile(
             title: Text('Offline'.i18n),
-            leading: const LeadingIcon(Icons.offline_pin, color: Color(0xff009a85)),
+            leading:
+                const LeadingIcon(Icons.offline_pin, color: Color(0xff009a85)),
           ),
           const QualityPicker('offline'),
           const FreezerDivider(),
           ListTile(
             title: Text('External downloads'.i18n),
-            leading: const LeadingIcon(Icons.file_download, color: Color(0xff2ba766)),
+            leading: const LeadingIcon(Icons.file_download,
+                color: Color(0xff2ba766)),
           ),
           const QualityPicker('download'),
         ],
@@ -531,87 +574,74 @@ class _QualityPickerState extends State<QualityPicker> {
 
   @override
   Widget build(BuildContext context) {
-    return Column(
-      children: <Widget>[
-        ListTile(
-          title: const Text('MP3 128kbps'),
-          leading: Radio(
-            groupValue: _quality,
+    return RadioGroup<AudioQuality>(
+      groupValue: _quality,
+      onChanged: (q) => _updateQuality(q!),
+      child: Column(
+        children: <Widget>[
+          const RadioListTile<AudioQuality>(
+            title: Text('MP3 128kbps'),
             value: AudioQuality.MP3_128,
-            onChanged: (q) => _updateQuality(q!),
           ),
-        ),
-        ListTile(
-          title: const Text('MP3 320kbps'),
-          leading: Radio(
-            groupValue: _quality,
+          const RadioListTile<AudioQuality>(
+            title: Text('MP3 320kbps'),
             value: AudioQuality.MP3_320,
-            onChanged: (q) => _updateQuality(q!),
           ),
-        ),
-        ListTile(
-          title: const Text('FLAC'),
-          leading: Radio(
-            groupValue: _quality,
+          const RadioListTile<AudioQuality>(
+            title: Text('FLAC'),
             value: AudioQuality.FLAC,
-            onChanged: (q) => _updateQuality(q!),
           ),
-        ),
-        if (widget.field == 'download')
-          ListTile(
+          if (widget.field == 'download')
+            RadioListTile<AudioQuality>(
               title: Text('Ask before downloading'.i18n),
-              leading: Radio(
-                groupValue: _quality,
-                value: AudioQuality.ASK,
-                onChanged: (q) => _updateQuality(q!),
-              ))
-      ],
+              value: AudioQuality.ASK,
+            )
+        ],
+      ),
     );
   }
 }
 
 class ContentLanguage {
-  String code;
-  String name;
-  ContentLanguage(this.code, this.name);
+  final String code;
+  final String name;
+  const ContentLanguage(this.code, this.name);
 
-  static List<ContentLanguage> get all => [
-        ContentLanguage('cs', 'Čeština'),
-        ContentLanguage('da', 'Dansk'),
-        ContentLanguage('de', 'Deutsch'),
-        ContentLanguage('en', 'English'),
-        ContentLanguage('us', 'English (us)'),
-        ContentLanguage('es', 'Español'),
-        ContentLanguage('mx', 'Español (latam)'),
-        ContentLanguage('fr', 'Français'),
-        ContentLanguage('hr', 'Hrvatski'),
-        ContentLanguage('id', 'Indonesia'),
-        ContentLanguage('it', 'Italiano'),
-        ContentLanguage('hu', 'Magyar'),
-        ContentLanguage('ms', 'Melayu'),
-        ContentLanguage('nl', 'Nederlands'),
-        ContentLanguage('no', 'Norsk'),
-        ContentLanguage('pl', 'Polski'),
-        ContentLanguage('br', 'Português (br)'),
-        ContentLanguage('pt', 'Português (pt)'),
-        ContentLanguage('ro', 'Română'),
-        ContentLanguage('sk', 'Slovenčina'),
-        ContentLanguage('sl', 'Slovenščina'),
-        ContentLanguage('sq', 'Shqip'),
-        ContentLanguage('sr', 'Srpski'),
-        ContentLanguage('fi', 'Suomi'),
-        ContentLanguage('sv', 'Svenska'),
-        ContentLanguage('tr', 'Türkçe'),
-        ContentLanguage('bg', 'Български'),
-        ContentLanguage('ru', 'Pусский'),
-        ContentLanguage('uk', 'Українська'),
-        ContentLanguage('he', 'עִברִית'),
-        ContentLanguage('ar', 'العربیة'),
-        ContentLanguage('cn', '中文'),
-        ContentLanguage('ja', '日本語'),
-        ContentLanguage('ko', '한국어'),
-        ContentLanguage('th', 'ภาษาไทย'),
-      ];
+  static List<ContentLanguage> all = const [
+    ContentLanguage('de', 'Deutsch'),
+    ContentLanguage('en', 'English'),
+    ContentLanguage('us', 'English (us)'),
+    ContentLanguage('es', 'Español'),
+    ContentLanguage('mx', 'Español (latam)'),
+    ContentLanguage('fr', 'Français'),
+    ContentLanguage('hr', 'Hrvatski'),
+    ContentLanguage('id', 'Indonesia'),
+    ContentLanguage('it', 'Italiano'),
+    ContentLanguage('hu', 'Magyar'),
+    ContentLanguage('ms', 'Melayu'),
+    ContentLanguage('nl', 'Nederlands'),
+    ContentLanguage('no', 'Norsk'),
+    ContentLanguage('pl', 'Polski'),
+    ContentLanguage('br', 'Português (br)'),
+    ContentLanguage('pt', 'Português (pt)'),
+    ContentLanguage('ro', 'Română'),
+    ContentLanguage('sk', 'Slovenčina'),
+    ContentLanguage('sl', 'Slovenščina'),
+    ContentLanguage('sq', 'Shqip'),
+    ContentLanguage('sr', 'Srpski'),
+    ContentLanguage('fi', 'Suomi'),
+    ContentLanguage('sv', 'Svenska'),
+    ContentLanguage('tr', 'Türkçe'),
+    ContentLanguage('bg', 'Български'),
+    ContentLanguage('ru', 'Pусский'),
+    ContentLanguage('uk', 'Українська'),
+    ContentLanguage('he', 'עִברִית'),
+    ContentLanguage('ar', 'العربیة'),
+    ContentLanguage('cn', '中文'),
+    ContentLanguage('ja', '日本語'),
+    ContentLanguage('ko', '한국어'),
+    ContentLanguage('th', 'ภาษาไทย'),
+  ];
 }
 
 class DeezerSettings extends StatefulWidget {
@@ -630,7 +660,8 @@ class _DeezerSettingsState extends State<DeezerSettings> {
         children: <Widget>[
           ListTile(
             title: Text('Content language'.i18n),
-            subtitle: Text('Not app language, used in headers. Now'.i18n + ': ${settings.deezerLanguage}'),
+            subtitle: Text('Not app language, used in headers. Now'.i18n +
+                ': ${settings.deezerLanguage}'),
             leading: const Icon(Icons.language),
             onTap: () {
               showDialog(
@@ -643,7 +674,8 @@ class _DeezerSettingsState extends State<DeezerSettings> {
                                   title: Text(ContentLanguage.all[i].name),
                                   subtitle: Text(ContentLanguage.all[i].code),
                                   onTap: () async {
-                                    setState(() => settings.deezerLanguage = ContentLanguage.all[i].code);
+                                    setState(() => settings.deezerLanguage =
+                                        ContentLanguage.all[i].code);
                                     await settings.save();
                                     if (context.mounted) {
                                       Navigator.of(context).pop();
@@ -655,7 +687,8 @@ class _DeezerSettingsState extends State<DeezerSettings> {
           ),
           ListTile(
             title: Text('Content country'.i18n),
-            subtitle: Text('Country used in headers. Now'.i18n + ': ${settings.deezerCountry}'),
+            subtitle: Text('Country used in headers. Now'.i18n +
+                ': ${settings.deezerCountry}'),
             leading: const Icon(Icons.vpn_lock),
             onTap: () {
               showDialog(
@@ -677,7 +710,8 @@ class _DeezerSettingsState extends State<DeezerSettings> {
                           ],
                         ),
                         onValuePicked: (Country country) {
-                          setState(() => settings.deezerCountry = country.isoCode ?? 'us');
+                          setState(() =>
+                              settings.deezerCountry = country.isoCode ?? 'us');
                           settings.save();
                         },
                       ));
@@ -685,7 +719,9 @@ class _DeezerSettingsState extends State<DeezerSettings> {
           ),
           ListTile(
             title: Text('Log tracks'.i18n),
-            subtitle: Text('Send track listen logs to Deezer, enable it for features like Flow to work properly'.i18n),
+            subtitle: Text(
+                'Send track listen logs to Deezer, enable it for features like Flow to work properly'
+                    .i18n),
             trailing: Switch(
               value: settings.logListen,
               onChanged: (bool v) {
@@ -780,10 +816,13 @@ class _FilenameTemplateDialogState extends State<FilenameTemplateDialog> {
         children: [
           TextField(
             decoration: InputDecoration(
-              floatingLabelStyle: TextStyle(color: Theme.of(context).primaryColor),
-                  focusedBorder: UnderlineInputBorder(
-                    borderSide: BorderSide(color: Theme.of(context).primaryColor), // Color of the underline when focused
-            ),
+              floatingLabelStyle:
+                  TextStyle(color: Theme.of(context).primaryColor),
+              focusedBorder: UnderlineInputBorder(
+                borderSide: BorderSide(
+                    color: Theme.of(context)
+                        .primaryColor), // Color of the underline when focused
+              ),
             ),
             cursorColor: Theme.of(context).primaryColor,
             controller: _controller,
@@ -793,7 +832,8 @@ class _FilenameTemplateDialogState extends State<FilenameTemplateDialog> {
           Text(
             'Valid variables are'.i18n +
                 ': %artists%, %artist%, %title%, %album%, %trackNumber%, %0trackNumber%, %feats%, %playlistTrackNumber%, %0playlistTrackNumber%, %year%, %date%\n\n' +
-                "If you want to use custom directory naming - use '/' as directory separator.".i18n,
+                "If you want to use custom directory naming - use '/' as directory separator."
+                    .i18n,
             style: const TextStyle(
               fontSize: 12.0,
             ),
@@ -802,35 +842,60 @@ class _FilenameTemplateDialogState extends State<FilenameTemplateDialog> {
       ),
       actions: [
         TextButton(
-                  style: ButtonStyle(
-          overlayColor: WidgetStateProperty.resolveWith<Color?>((Set<WidgetState> states) {if (states.contains(WidgetState.pressed)) {return Theme.of(context).primaryColor.withOpacity(0.3);}return null;}),
-         ),
+          style: ButtonStyle(
+            overlayColor: material.WidgetStateProperty.resolveWith<Color?>(
+                (Set<material.WidgetState> states) {
+              if (states.contains(material.WidgetState.pressed)) {
+                return Theme.of(context).primaryColor.withValues(alpha: 0.3);
+              }
+              return null;
+            }),
+          ),
           child: Text('Cancel'.i18n),
           onPressed: () => Navigator.of(context).pop(),
         ),
         TextButton(
-                  style: ButtonStyle(
-          overlayColor: WidgetStateProperty.resolveWith<Color?>((Set<WidgetState> states) {if (states.contains(WidgetState.pressed)) {return Theme.of(context).primaryColor.withOpacity(0.3);}return null;}),
-         ),
+          style: ButtonStyle(
+            overlayColor: material.WidgetStateProperty.resolveWith<Color?>(
+                (Set<material.WidgetState> states) {
+              if (states.contains(material.WidgetState.pressed)) {
+                return Theme.of(context).primaryColor.withValues(alpha: 0.3);
+              }
+              return null;
+            }),
+          ),
           child: Text('Reset'.i18n),
           onPressed: () {
-            _controller.value = _controller.value.copyWith(text: '%artist% - %title%');
+            _controller.value =
+                _controller.value.copyWith(text: '%artist% - %title%');
             _new = '%artist% - %title%';
           },
         ),
         TextButton(
-                                                      style: ButtonStyle(
-                                              overlayColor: WidgetStateProperty.resolveWith<Color?>((Set<WidgetState> states) {if (states.contains(WidgetState.pressed)) {return Theme.of(context).primaryColor.withOpacity(0.3);}return null;}),
-                                            ),
+          style: ButtonStyle(
+            overlayColor: material.WidgetStateProperty.resolveWith<Color?>(
+                (Set<material.WidgetState> states) {
+              if (states.contains(material.WidgetState.pressed)) {
+                return Theme.of(context).primaryColor.withValues(alpha: 0.3);
+              }
+              return null;
+            }),
+          ),
           child: Text('Clear'.i18n),
           onPressed: () => _controller.clear(),
         ),
         TextButton(
-                                                      style: ButtonStyle(
-                                              overlayColor: WidgetStateProperty.resolveWith<Color?>((Set<WidgetState> states) {if (states.contains(WidgetState.pressed)) {return Theme.of(context).primaryColor.withOpacity(0.3);}return null;}),
-                                            ),
+          style: ButtonStyle(
+            overlayColor: material.WidgetStateProperty.resolveWith<Color?>(
+                (Set<material.WidgetState> states) {
+              if (states.contains(material.WidgetState.pressed)) {
+                return Theme.of(context).primaryColor.withValues(alpha: 0.3);
+              }
+              return null;
+            }),
+          ),
           child: Text('Save'.i18n),
-          onPressed: () async {
+          onPressed: () {
             widget.onSave(_new);
             Navigator.of(context).pop();
           },
@@ -849,7 +914,8 @@ class DownloadsSettings extends StatefulWidget {
 
 class _DownloadsSettingsState extends State<DownloadsSettings> {
   double _downloadThreads = settings.downloadThreads.toDouble();
-  final TextEditingController _artistSeparatorController = TextEditingController(text: settings.artistSeparator);
+  final TextEditingController _artistSeparatorController =
+      TextEditingController(text: settings.artistSeparator);
 
   @override
   Widget build(BuildContext context) {
@@ -893,7 +959,8 @@ class _DownloadsSettingsState extends State<DownloadsSettings> {
               showDialog(
                   context: context,
                   builder: (context) {
-                    return FilenameTemplateDialog(settings.downloadFilename, (f) async {
+                    return FilenameTemplateDialog(settings.downloadFilename,
+                        (f) async {
                       setState(() => settings.downloadFilename = f);
                       await settings.save();
                     });
@@ -902,13 +969,15 @@ class _DownloadsSettingsState extends State<DownloadsSettings> {
           ),
           ListTile(
             title: Text('Singleton naming'.i18n),
-            subtitle: Text('Currently'.i18n + ': ${settings.singletonFilename}'),
+            subtitle:
+                Text('Currently'.i18n + ': ${settings.singletonFilename}'),
             leading: const Icon(Icons.text_format),
             onTap: () {
               showDialog(
                   context: context,
                   builder: (context) {
-                    return FilenameTemplateDialog(settings.singletonFilename, (f) async {
+                    return FilenameTemplateDialog(settings.singletonFilename,
+                        (f) async {
                       setState(() => settings.singletonFilename = f);
                       await settings.save();
                     });
@@ -918,7 +987,8 @@ class _DownloadsSettingsState extends State<DownloadsSettings> {
           Padding(
             padding: const EdgeInsets.symmetric(horizontal: 16.0),
             child: Text(
-              'Download threads'.i18n + ': ${_downloadThreads.round().toString()}',
+              'Download threads'.i18n +
+                  ': ${_downloadThreads.round().toString()}',
               style: const TextStyle(fontSize: 16.0),
             ),
           ),
@@ -938,19 +1008,32 @@ class _DownloadsSettingsState extends State<DownloadsSettings> {
                 await settings.save();
 
                 //Prevent null
-                if (val > 8 && cache.threadsWarning != true && context.mounted) {
+                if (val > 8 &&
+                    cache.threadsWarning != true &&
+                    context.mounted) {
                   showDialog(
                       context: context,
                       builder: (context) {
                         return AlertDialog(
                           title: Text('Warning'.i18n),
                           content: Text(
-                              'Using too many concurrent downloads on older/weaker devices might cause crashes!'.i18n),
+                              'Using too many concurrent downloads on older/weaker devices might cause crashes!'
+                                  .i18n),
                           actions: [
                             TextButton(
-                                                                          style: ButtonStyle(
-                                              overlayColor: WidgetStateProperty.resolveWith<Color?>((Set<WidgetState> states) {if (states.contains(WidgetState.pressed)) {return Theme.of(context).primaryColor.withOpacity(0.3);}return null;}),
-                                            ),
+                              style: ButtonStyle(
+                                overlayColor: material.WidgetStateProperty
+                                    .resolveWith<Color?>(
+                                        (Set<material.WidgetState> states) {
+                                  if (states
+                                      .contains(material.WidgetState.pressed)) {
+                                    return Theme.of(context)
+                                        .primaryColor
+                                        .withValues(alpha: 0.3);
+                                  }
+                                  return null;
+                                }),
+                              ),
                               child: Text('Dismiss'.i18n),
                               onPressed: () => Navigator.of(context).pop(),
                             )
@@ -966,8 +1049,8 @@ class _DownloadsSettingsState extends State<DownloadsSettings> {
           ListTile(
             title: Text('Tags'.i18n),
             leading: const Icon(Icons.label),
-            onTap: () =>
-                Navigator.of(context).push(MaterialPageRoute(builder: (context) => const TagSelectionScreen())),
+            onTap: () => Navigator.of(context).push(MaterialPageRoute(
+                builder: (context) => const TagSelectionScreen())),
           ),
           ListTile(
             title: Text('Create folders for artist'.i18n),
@@ -1054,17 +1137,20 @@ class _DownloadsSettingsState extends State<DownloadsSettings> {
               leading: const Icon(Icons.image)),
           ListTile(
               title: Text('Album cover resolution'.i18n),
-              subtitle: Text("WARNING: Resolutions above 1200 aren't officially supported".i18n),
+              subtitle: Text(
+                  "WARNING: Resolutions above 1200 aren't officially supported"
+                      .i18n),
               leading: const Icon(Icons.image),
               trailing: SizedBox(
                   width: 75.0,
                   child: DropdownButton<int>(
                     value: settings.albumArtResolution,
                     items: [400, 800, 1000, 1200, 1400, 1600, 1800]
-                        .map<DropdownMenuItem<int>>((int i) => DropdownMenuItem<int>(
-                              value: i,
-                              child: Text(i.toString()),
-                            ))
+                        .map<DropdownMenuItem<int>>(
+                            (int i) => DropdownMenuItem<int>(
+                                  value: i,
+                                  child: Text(i.toString()),
+                                ))
                         .toList(),
                     onChanged: (int? n) async {
                       setState(() {
@@ -1075,7 +1161,8 @@ class _DownloadsSettingsState extends State<DownloadsSettings> {
                   ))),
           ListTile(
               title: Text('Create .nomedia files'.i18n),
-              subtitle: Text('To prevent gallery being filled with album art'.i18n),
+              subtitle:
+                  Text('To prevent gallery being filled with album art'.i18n),
               trailing: Switch(
                 value: settings.nomediaFiles,
                 onChanged: (v) {
@@ -1096,12 +1183,15 @@ class _DownloadsSettingsState extends State<DownloadsSettings> {
                   settings.artistSeparator = s;
                   await settings.save();
                 },
-                            decoration: InputDecoration(
-                              floatingLabelStyle: TextStyle(color: Theme.of(context).primaryColor),
+                decoration: InputDecoration(
+                  floatingLabelStyle:
+                      TextStyle(color: Theme.of(context).primaryColor),
                   focusedBorder: UnderlineInputBorder(
-                    borderSide: BorderSide(color: Theme.of(context).primaryColor), // Color of the underline when focused
-            ),
-                            ),
+                    borderSide: BorderSide(
+                        color: Theme.of(context)
+                            .primaryColor), // Color of the underline when focused
+                  ),
+                ),
               ),
             ),
           ),
@@ -1109,7 +1199,8 @@ class _DownloadsSettingsState extends State<DownloadsSettings> {
           ListTile(
             title: Text('Download Log'.i18n),
             leading: const Icon(Icons.sticky_note_2),
-            onTap: () => Navigator.of(context).push(MaterialPageRoute(builder: (context) => const DownloadLogViewer())),
+            onTap: () => Navigator.of(context).push(MaterialPageRoute(
+                builder: (context) => const DownloadLogViewer())),
           )
         ],
       ),
@@ -1210,7 +1301,9 @@ class _GeneralSettingsState extends State<GeneralSettings> {
                           setState(() => settings.offlineMode = false);
                         } else {
                           Fluttertoast.showToast(
-                              msg: 'Error logging in, check your internet connections.'.i18n,
+                              msg:
+                                  'Error logging in, check your internet connections.'
+                                      .i18n,
                               gravity: ToastGravity.BOTTOM,
                               toastLength: Toast.LENGTH_SHORT);
                         }
@@ -1231,7 +1324,8 @@ class _GeneralSettingsState extends State<GeneralSettings> {
           ),
           ListTile(
             title: Text('Copy ARL'.i18n),
-            subtitle: Text('Copy userToken/ARL Cookie for use in other apps.'.i18n),
+            subtitle:
+                Text('Copy userToken/ARL Cookie for use in other apps.'.i18n),
             leading: const Icon(Icons.lock),
             onTap: () async {
               await FlutterClipboard.copy(settings.arl ?? '');
@@ -1242,7 +1336,9 @@ class _GeneralSettingsState extends State<GeneralSettings> {
           ),
           ListTile(
             title: Text('Enable equalizer'.i18n),
-            subtitle: Text('Might enable some equalizer apps to work. Requires restart of ReFreezer'.i18n),
+            subtitle: Text(
+                'Might enable some equalizer apps to work. Requires restart of ReFreezer'
+                    .i18n),
             leading: const Icon(Icons.equalizer),
             trailing: Switch(
               value: settings.enableEqualizer,
@@ -1254,7 +1350,9 @@ class _GeneralSettingsState extends State<GeneralSettings> {
           ),
           ListTile(
             title: Text('LastFM'.i18n),
-            subtitle: Text((settings.lastFMUsername != null) ? 'Log out'.i18n : 'Login to enable scrobbling.'.i18n),
+            subtitle: Text((settings.lastFMUsername != null)
+                ? 'Log out'.i18n
+                : 'Login to enable scrobbling.'.i18n),
             leading: const Icon(FontAwesome5.lastfm),
             onTap: () async {
               if (settings.lastFMUsername != null) {
@@ -1289,10 +1387,13 @@ class _GeneralSettingsState extends State<GeneralSettings> {
                   settings.lastFMAPIKey = s;
                   await settings.save();
                 },
-                  decoration: InputDecoration(
-                    floatingLabelStyle: TextStyle(color: Theme.of(context).primaryColor),
+                decoration: InputDecoration(
+                  floatingLabelStyle:
+                      TextStyle(color: Theme.of(context).primaryColor),
                   focusedBorder: UnderlineInputBorder(
-                    borderSide: BorderSide(color: Theme.of(context).primaryColor), // Color of the underline when focused
+                    borderSide: BorderSide(
+                        color: Theme.of(context)
+                            .primaryColor), // Color of the underline when focused
                   ),
                 ),
               ),
@@ -1309,10 +1410,13 @@ class _GeneralSettingsState extends State<GeneralSettings> {
                   settings.lastFMAPISecret = s;
                   await settings.save();
                 },
-                  decoration: InputDecoration(
-                    floatingLabelStyle: TextStyle(color: Theme.of(context).primaryColor),
+                decoration: InputDecoration(
+                  floatingLabelStyle:
+                      TextStyle(color: Theme.of(context).primaryColor),
                   focusedBorder: UnderlineInputBorder(
-                    borderSide: BorderSide(color: Theme.of(context).primaryColor), // Color of the underline when focused
+                    borderSide: BorderSide(
+                        color: Theme.of(context)
+                            .primaryColor), // Color of the underline when focused
                   ),
                 ),
               ),
@@ -1345,8 +1449,8 @@ class _GeneralSettingsState extends State<GeneralSettings> {
           ListTile(
             title: Text('Application Log'.i18n),
             leading: const Icon(Icons.sticky_note_2),
-            onTap: () =>
-                Navigator.of(context).push(MaterialPageRoute(builder: (context) => const ApplicationLogViewer())),
+            onTap: () => Navigator.of(context).push(MaterialPageRoute(
+                builder: (context) => const ApplicationLogViewer())),
           ),
           const FreezerDivider(),
           ListTile(
@@ -1367,16 +1471,36 @@ class _GeneralSettingsState extends State<GeneralSettings> {
                         content: Text('Are you sure you want to log out?'.i18n),
                         actions: <Widget>[
                           TextButton(
-                                                      style: ButtonStyle(
-          overlayColor: WidgetStateProperty.resolveWith<Color?>((Set<WidgetState> states) {if (states.contains(WidgetState.pressed)) {return Theme.of(context).primaryColor.withOpacity(0.3);}return null;}),
-         ),
+                            style: ButtonStyle(
+                              overlayColor: material.WidgetStateProperty
+                                  .resolveWith<Color?>(
+                                      (Set<material.WidgetState> states) {
+                                if (states
+                                    .contains(material.WidgetState.pressed)) {
+                                  return Theme.of(context)
+                                      .primaryColor
+                                      .withValues(alpha: 0.3);
+                                }
+                                return null;
+                              }),
+                            ),
                             child: Text('Cancel'.i18n),
                             onPressed: () => Navigator.of(context).pop(),
                           ),
                           TextButton(
-                                                      style: ButtonStyle(
-          overlayColor: WidgetStateProperty.resolveWith<Color?>((Set<WidgetState> states) {if (states.contains(WidgetState.pressed)) {return Theme.of(context).primaryColor.withOpacity(0.3);}return null;}),
-         ),
+                            style: ButtonStyle(
+                              overlayColor: material.WidgetStateProperty
+                                  .resolveWith<Color?>(
+                                      (Set<material.WidgetState> states) {
+                                if (states
+                                    .contains(material.WidgetState.pressed)) {
+                                  return Theme.of(context)
+                                      .primaryColor
+                                      .withValues(alpha: 0.3);
+                                }
+                                return null;
+                              }),
+                            ),
                             //child: Text('(ARL ONLY) Continue'.i18n),
                             child: Text('Continue'.i18n),
                             onPressed: () async {
@@ -1430,11 +1554,13 @@ class _LastFMLoginState extends State<LastFMLogin> {
         children: [
           TextField(
             cursorColor: Theme.of(context).primaryColor,
-            decoration: InputDecoration(hintText: 'Username'.i18n,
-            floatingLabelStyle: TextStyle(color: Theme.of(context).primaryColor),
-                              focusedBorder: UnderlineInputBorder(
-                    borderSide: BorderSide(color: Theme.of(context).primaryColor),
-                              ),
+            decoration: InputDecoration(
+              hintText: 'Username'.i18n,
+              floatingLabelStyle:
+                  TextStyle(color: Theme.of(context).primaryColor),
+              focusedBorder: UnderlineInputBorder(
+                borderSide: BorderSide(color: Theme.of(context).primaryColor),
+              ),
             ),
             onChanged: (v) => _username = v,
           ),
@@ -1442,11 +1568,13 @@ class _LastFMLoginState extends State<LastFMLogin> {
           TextField(
             cursorColor: Theme.of(context).primaryColor,
             obscureText: true,
-            decoration: InputDecoration(hintText: 'Password'.i18n,
-            floatingLabelStyle: TextStyle(color: Theme.of(context).primaryColor),
-                                          focusedBorder: UnderlineInputBorder(
-                    borderSide: BorderSide(color: Theme.of(context).primaryColor),
-                              ),
+            decoration: InputDecoration(
+              hintText: 'Password'.i18n,
+              floatingLabelStyle:
+                  TextStyle(color: Theme.of(context).primaryColor),
+              focusedBorder: UnderlineInputBorder(
+                borderSide: BorderSide(color: Theme.of(context).primaryColor),
+              ),
             ),
             onChanged: (v) => _password = v,
           )
@@ -1454,16 +1582,28 @@ class _LastFMLoginState extends State<LastFMLogin> {
       ),
       actions: [
         TextButton(
-                                    style: ButtonStyle(
-          overlayColor: WidgetStateProperty.resolveWith<Color?>((Set<WidgetState> states) {if (states.contains(WidgetState.pressed)) {return Theme.of(context).primaryColor.withOpacity(0.3);}return null;}),
-         ),
+          style: ButtonStyle(
+            overlayColor: material.WidgetStateProperty.resolveWith<Color?>(
+                (Set<material.WidgetState> states) {
+              if (states.contains(material.WidgetState.pressed)) {
+                return Theme.of(context).primaryColor.withValues(alpha: 0.3);
+              }
+              return null;
+            }),
+          ),
           child: Text('Cancel'.i18n),
           onPressed: () => Navigator.of(context).pop(),
         ),
         TextButton(
-                                    style: ButtonStyle(
-          overlayColor: WidgetStateProperty.resolveWith<Color?>((Set<WidgetState> states) {if (states.contains(WidgetState.pressed)) {return Theme.of(context).primaryColor.withOpacity(0.3);}return null;}),
-         ),
+          style: ButtonStyle(
+            overlayColor: material.WidgetStateProperty.resolveWith<Color?>(
+                (Set<material.WidgetState> states) {
+              if (states.contains(material.WidgetState.pressed)) {
+                return Theme.of(context).primaryColor.withValues(alpha: 0.3);
+              }
+              return null;
+            }),
+          ),
           child: Text('Login'.i18n),
           onPressed: () async {
             LastFM last;
@@ -1496,23 +1636,30 @@ class StorageInfo {
   final String appFilesDir;
   final int availableBytes;
 
-  StorageInfo({required this.rootDir, required this.appFilesDir, required this.availableBytes});
+  StorageInfo(
+      {required this.rootDir,
+      required this.appFilesDir,
+      required this.availableBytes});
 }
 
 Future<List<StorageInfo>> getStorageInfo() async {
-  final externalDirectories = await ExternalPath.getExternalStorageDirectories();
+  final externalDirectories =
+      await ExternalPath.getExternalStorageDirectories();
 
   List<StorageInfo> storageInfoList = [];
 
   if (externalDirectories.isNotEmpty) {
     for (var dir in externalDirectories) {
-      var availableMegaBytes = (await DiskSpacePlus.getFreeDiskSpaceForPath(dir)) ?? 0.0;
+      var availableMegaBytes =
+          (await DiskSpacePlus.getFreeDiskSpaceForPath(dir)) ?? 0.0;
 
       storageInfoList.add(
         StorageInfo(
           rootDir: dir,
           appFilesDir: dir,
-          availableBytes: availableMegaBytes > 0 ? (availableMegaBytes * 1000000).floor() : 0,
+          availableBytes: availableMegaBytes > 0
+              ? (availableMegaBytes * 1000000).floor()
+              : 0,
         ),
       );
     }
@@ -1581,11 +1728,14 @@ class _DirectoryPickerState extends State<DirectoryPicker> {
                             if (snapshot.hasError) return const ErrorScreen();
                             if (!snapshot.hasData) {
                               return Padding(
-                                padding: const EdgeInsets.symmetric(vertical: 8.0),
+                                padding:
+                                    const EdgeInsets.symmetric(vertical: 8.0),
                                 child: Row(
                                   mainAxisAlignment: MainAxisAlignment.center,
                                   children: <Widget>[
-                                    CircularProgressIndicator(color: Theme.of(context).primaryColor,)
+                                    CircularProgressIndicator(
+                                      color: Theme.of(context).primaryColor,
+                                    )
                                   ],
                                 ),
                               );
@@ -1593,7 +1743,8 @@ class _DirectoryPickerState extends State<DirectoryPicker> {
                             return Column(
                               mainAxisSize: MainAxisSize.min,
                               children: <Widget>[
-                                ...List.generate(snapshot.data?.length ?? 0, (i) {
+                                ...List.generate(snapshot.data?.length ?? 0,
+                                    (i) {
                                   StorageInfo si = snapshot.data![i];
                                   return ListTile(
                                     title: Text(si.rootDir),
@@ -1641,7 +1792,9 @@ class _DirectoryPickerState extends State<DirectoryPicker> {
           }
           if (!snapshot.hasData) {
             return Center(
-              child: CircularProgressIndicator(color: Theme.of(context).primaryColor,),
+              child: CircularProgressIndicator(
+                color: Theme.of(context).primaryColor,
+              ),
             );
           }
 
@@ -1657,7 +1810,9 @@ class _DirectoryPickerState extends State<DirectoryPicker> {
                 onTap: () {
                   setState(() {
                     if (_root == _path) {
-                      Fluttertoast.showToast(msg: 'Permission denied'.i18n, gravity: ToastGravity.BOTTOM);
+                      Fluttertoast.showToast(
+                          msg: 'Permission denied'.i18n,
+                          gravity: ToastGravity.BOTTOM);
                       return;
                     }
                     _previous = _path;
@@ -1721,7 +1876,6 @@ class _CreditsScreenState extends State<CreditsScreen> {
   //   ['RTWO2', 'Persian']
   // ];
 
-
   @override
   void initState() {
     PackageInfo.fromPlatform().then((info) {
@@ -1748,7 +1902,8 @@ class _CreditsScreenState extends State<CreditsScreen> {
           ListTile(
             title: Text('Telegram Channel'.i18n),
             subtitle: Text('To get latest releases'.i18n),
-            leading: const Icon(FontAwesome5.telegram, color: Color(0xFF27A2DF), size: 36.0),
+            leading: const Icon(FontAwesome5.telegram,
+                color: Color(0xFF27A2DF), size: 36.0),
             onTap: () {
               launchUrlString('https://t.me/SaturnReleases');
             },
@@ -1756,7 +1911,8 @@ class _CreditsScreenState extends State<CreditsScreen> {
           ListTile(
             title: Text('Telegram Group'.i18n),
             subtitle: Text('Official chat'.i18n),
-            leading: const Icon(FontAwesome5.telegram, color: Colors.cyan, size: 36.0),
+            leading: const Icon(FontAwesome5.telegram,
+                color: Colors.cyan, size: 36.0),
             onTap: () {
               launchUrlString('https://t.me/SaturnDiscuss');
             },
@@ -1764,7 +1920,8 @@ class _CreditsScreenState extends State<CreditsScreen> {
           ListTile(
             title: Text('Discord'.i18n),
             subtitle: Text('Official Discord server'.i18n),
-            leading: const Icon(FontAwesome5.discord, color: Color(0xff7289da), size: 36.0),
+            leading: const Icon(FontAwesome5.discord,
+                color: Color(0xff7289da), size: 36.0),
             onTap: () {
               launchUrlString('https://saturn.kim/discord');
             },
@@ -1779,64 +1936,65 @@ class _CreditsScreenState extends State<CreditsScreen> {
           ),
           ListTile(
             title: const Text('Donate'),
-            subtitle: const Text('Send crypto to the Saturn fund to support the development.'),
-            leading: const Icon(FontAwesome5.bitcoin, color: Color.fromRGBO(247,147,26, 58), size: 36.0),
+            subtitle: const Text(
+                'Send crypto to the Saturn fund to support the development.'),
+            leading: const Icon(FontAwesome5.bitcoin,
+                color: Color.fromRGBO(247, 147, 26, 58), size: 36.0),
             onTap: () {
               launchUrlString('https://fund.saturn.kim/');
             },
           ),
           const FreezerDivider(),
           ListTile(
-            title: const Text('bw86'),
-            subtitle: const Text('Logo Designer, Developer'),
-            onTap: () {
-                                  showDialog(
-                                    context: context,
-                                    builder: (context) {
-                                      return AlertDialog(
-                                        title: Text('Visit my site!'.i18n),
-                                        content: Text('www.autow.in'.i18n),
-                                        actions: <Widget>[
-                                          TextButton(
-                                            style: ButtonStyle(
-                                              foregroundColor: WidgetStateProperty.all<Color>(Theme.of(context).primaryColor)
-                                            ),
-                                            onPressed: () => Navigator.of(context).pop(),
-                                            child: Text('AGREE'.i18n),
-                                          ),
-                                        ],
-                                      );
-                                    },
-                                  );
-            }
-          ),
+              title: const Text('bw86'),
+              subtitle: const Text('Logo Designer, Developer'),
+              onTap: () {
+                showDialog(
+                  context: context,
+                  builder: (context) {
+                    return AlertDialog(
+                      title: Text('Visit my site!'.i18n),
+                      content: Text('www.autow.in'.i18n),
+                      actions: <Widget>[
+                        TextButton(
+                          style: ButtonStyle(
+                              foregroundColor: WidgetStateProperty.all<Color>(
+                                  Theme.of(context).primaryColor)),
+                          onPressed: () => Navigator.of(context).pop(),
+                          child: Text('AGREE'.i18n),
+                        ),
+                      ],
+                    );
+                  },
+                );
+              }),
           ListTile(
-            title: const Text('Matt'),
-            subtitle: const Text('Developer'),
-                        onTap: () {
-                                  showDialog(
-                                    context: context,
-                                    builder: (context) {
-                                      return AlertDialog(
-                                        title: Text('hiya!! please star the repo'.i18n),
-                                        content: Text('github.com/Sheathed'.i18n),
-                                        actions: <Widget>[
-                                          TextButton(
-                                            style: ButtonStyle(
-                                              foregroundColor: WidgetStateProperty.all<Color>(Theme.of(context).primaryColor)
-                                            ),
-                                            onPressed: () => Navigator.of(context).pop(),
-                                            child: Text('AGREE'.i18n),
-                                          ),
-                                        ],
-                                      );
-                                    },
-                                  );
-            }
-          ),
+              title: const Text('Matt'),
+              subtitle: const Text('Developer'),
+              onTap: () {
+                showDialog(
+                  context: context,
+                  builder: (context) {
+                    return AlertDialog(
+                      title: Text('hiya!! please star the repo'.i18n),
+                      content: Text('github.com/Sheathed'.i18n),
+                      actions: <Widget>[
+                        TextButton(
+                          style: ButtonStyle(
+                              foregroundColor: WidgetStateProperty.all<Color>(
+                                  Theme.of(context).primaryColor)),
+                          onPressed: () => Navigator.of(context).pop(),
+                          child: Text('AGREE'.i18n),
+                        ),
+                      ],
+                    );
+                  },
+                );
+              }),
           const ListTile(
             title: Text('DJDoubleD'),
-            subtitle: Text('For allowing us to use his updated source & maintained forks of discontinued libs.'),
+            subtitle: Text(
+                'For allowing us to use his updated source & maintained forks of discontinued libs.'),
           ),
           const ListTile(
             title: Text('ettex, Xander Null, Francesco, Tobs'),
@@ -1873,16 +2031,38 @@ class LicensesScreen extends StatefulWidget {
 }
 
 class _LicensesScreenState extends State<LicensesScreen> {
-
-static final List<List<String>> licenses = [
-  ['Scrobblenaut', 'NPL | DJDoubleD & Nebulino', 'https://github.com/DJDoubleD/Scrobblenaut'],
-  ['move_to_background', 'MIT | DJDoubleD & Coin-ai', 'https://github.com/DJDoubleD/move_to_background'],
-  ['marquee', 'MIT | DJDoubleD & MarcelGarus', 'https://github.com/DJDoubleD/marquee'],
-  ['external_path', 'MIT | DJDoubleD & Siruss187', 'https://github.com/DJDoubleD/external_path'],
-  ['equalizer_flutter', 'MIT | DJDoubleD & nickwph', 'https://github.com/DJDoubleD/equalizer_flutter'],
-  ['custom_navigator', 'MIT | DJDoubleD & justprodev', 'https://github.com/DJDoubleD/custom_navigator'],
-];
-
+  static final List<List<String>> licenses = [
+    [
+      'Scrobblenaut',
+      'NPL | DJDoubleD & Nebulino',
+      'https://github.com/DJDoubleD/Scrobblenaut'
+    ],
+    [
+      'move_to_background',
+      'MIT | DJDoubleD & Coin-ai',
+      'https://github.com/DJDoubleD/move_to_background'
+    ],
+    [
+      'marquee',
+      'MIT | DJDoubleD & MarcelGarus',
+      'https://github.com/DJDoubleD/marquee'
+    ],
+    [
+      'external_path',
+      'MIT | DJDoubleD & Siruss187',
+      'https://github.com/DJDoubleD/external_path'
+    ],
+    [
+      'equalizer_flutter',
+      'MIT | DJDoubleD & nickwph',
+      'https://github.com/DJDoubleD/equalizer_flutter'
+    ],
+    [
+      'custom_navigator',
+      'MIT | DJDoubleD & justprodev',
+      'https://github.com/DJDoubleD/custom_navigator'
+    ],
+  ];
 
   @override
   void initState() {
@@ -1901,7 +2081,7 @@ static final List<List<String>> licenses = [
                     title: Text(licenses[i][0]),
                     subtitle: Text(licenses[i][1] + ' | Click to view Repo'),
                     onTap: () {
-                    launchUrlString(licenses[i][2]);
+                      launchUrlString(licenses[i][2]);
                     },
                   )),
           Padding(
