@@ -6,32 +6,36 @@ part of 'cache.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Cache _$CacheFromJson(Map<String, dynamic> json) => Cache(
-      libraryTracks: (json['libraryTracks'] as List<dynamic>?)
-          ?.map((e) => e as String)
-          .toList(),
-    )
-      ..history = (json['history'] as List<dynamic>?)
+Cache _$CacheFromJson(Map<String, dynamic> json) =>
+    Cache(
+        libraryTracks: (json['libraryTracks'] as List<dynamic>?)
+            ?.map((e) => e as String)
+            .toList(),
+      )
+      ..history =
+          (json['history'] as List<dynamic>?)
               ?.map((e) => Track.fromJson(e as Map<String, dynamic>))
               .toList() ??
           []
-      ..sorts = (json['sorts'] as List<dynamic>?)
+      ..sorts =
+          (json['sorts'] as List<dynamic>?)
               ?.map((e) => Sorting.fromJson(e as Map<String, dynamic>))
               .toList() ??
           []
-      ..searchHistory =
-          Cache._searchHistoryFromJson(json['searchHistory2'] as List?)
+      ..searchHistory = Cache._searchHistoryFromJson(
+        json['searchHistory2'] as List?,
+      )
       ..threadsWarning = json['threadsWarning'] as bool? ?? false
       ..lastUpdateCheck = (json['lastUpdateCheck'] as num?)?.toInt() ?? 0;
 
 Map<String, dynamic> _$CacheToJson(Cache instance) => <String, dynamic>{
-      'libraryTracks': instance.libraryTracks,
-      'history': instance.history,
-      'sorts': instance.sorts,
-      'searchHistory2': Cache._searchHistoryToJson(instance.searchHistory),
-      'threadsWarning': instance.threadsWarning,
-      'lastUpdateCheck': instance.lastUpdateCheck,
-    };
+  'libraryTracks': instance.libraryTracks,
+  'history': instance.history,
+  'sorts': instance.sorts,
+  'searchHistory2': Cache._searchHistoryToJson(instance.searchHistory),
+  'threadsWarning': instance.threadsWarning,
+  'lastUpdateCheck': instance.lastUpdateCheck,
+};
 
 SearchHistoryItem _$SearchHistoryItemFromJson(Map<String, dynamic> json) =>
     SearchHistoryItem(
