@@ -26,21 +26,16 @@ import 'ui/home_screen.dart';
 import 'ui/library.dart';
 import 'ui/login_screen.dart';
 import 'ui/player_bar.dart';
-import 'ui/updater.dart';
 import 'ui/search.dart';
 import 'utils/logging.dart';
 import 'utils/navigator_keys.dart';
 import 'ui/restartable.dart';
-import 'ui/clubs_screen.dart';
-import 'api/clubs.dart';
 
 late Function updateTheme;
 late Function logOut;
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  getIt.registerSingleton<dataMGMT>(dataMGMT());
-  getIt.registerSingleton<SocketManagement>(SocketManagement(address: 'https://clubs.saturn.kim:443', clubRoom: ClubRoom()),);
 
   await Permission.notification.isDenied.then((value) {
     if (value) {
@@ -491,7 +486,9 @@ class _MainScreenState extends State<MainScreen>
           // While audio_service is initializing
           return Scaffold(
             body: Center(
-              child: CircularProgressIndicator(color: Theme.of(context).primaryColor,),
+              child: CircularProgressIndicator(
+                color: Theme.of(context).primaryColor,
+              ),
             ),
           );
         }
